Finds the effect sizes for all genes in the original dataset, regardless of significance.

Finds the effect sizes for all genes in the original dataset, regardless of significance.

calcEffectSizes(countMatrix, condition)

Arguments

countMatrix
Matrix. A simulated counts matrix, sans labels.

condition
Factor. The condition labels for the simulated cells. If more than 2 conditions are given, the first will be compared to all others by default.

Value

A vector of cohen's d effect sizes for each gene.

Examples

data("mouseBrainSubsetSCE")
res <- calcEffectSizes(assay(mouseBrainSubsetSCE, "counts"),
                       condition = colData(mouseBrainSubsetSCE)[, "level1class"])