convertSCEToSeurat Converts sce object to seurat while retaining all assays and metadata

convertSCEToSeurat Converts sce object to seurat while retaining all assays and metadata

convertSCEToSeurat(
  inSCE,
  countsAssay = NULL,
  normAssay = NULL,
  scaledAssay = NULL,
  copyColData = FALSE,
  copyReducedDim = FALSE,
  copyDecontX = FALSE,
  pcaReducedDim = NULL,
  icaReducedDim = NULL,
  tsneReducedDim = NULL,
  umapReducedDim = NULL
)

Arguments

inSCE
A SingleCellExperiment object to convert to a Seurat object.

countsAssay
Which assay to use from sce object for raw counts. Default NULL.

normAssay
Which assay to use from sce object for normalized data. Default NULL.

scaledAssay
Which assay to use from sce object for scaled data. Default NULL.

copyColData
Boolean. Whether copy 'colData' of SCE object to the 'meta.data' of Seurat object. Default FALSE.

copyReducedDim
Boolean. Whether copy 'reducedDims' of the SCE object to the 'reductions' of Seurat object. Default FALSE.

copyDecontX
Boolean. Whether copy 'decontXcounts' assay of the SCE object to the 'assays' of Seurat object. Default TRUE.

pcaReducedDim
Specify a character value indicating the name of the reducedDim to store as default pca computation in the output seurat object. Default is NULL which will not store any reducedDim as the default pca. This will only work when copyReducedDim parameter is set to TRUE.

icaReducedDim
Specify a character value indicating the name of the reducedDim to store as default ica computation in the output seurat object. Default is NULL which will not store any reducedDim as the default ica. This will only work when copyReducedDim parameter is set to TRUE.

tsneReducedDim
Specify a character value indicating the name of the reducedDim to store as default tsne computation in the output seurat object. Default is NULL which will not store any reducedDim as the default tsne. This will only work when copyReducedDim parameter is set to TRUE.

umapReducedDim
Specify a character value indicating the name of the reducedDim to store as default umap computation in the output seurat object. Default is NULL which will not store any reducedDim as the default umap. This will only work when copyReducedDim parameter is set to TRUE.

Value

Updated seurat object that contains all data from the input sce object

Examples

data(scExample, package = "singleCellTK")
seurat <- convertSCEToSeurat(sce)

#> Warning: replacing previous import 'spatstat.utils::RelevantNA' by 'spatstat.sparse::RelevantNA' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::grokIndexVector' by 'spatstat.sparse::grokIndexVector' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::fullIndexSequence' by 'spatstat.sparse::fullIndexSequence' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::positiveIndex' by 'spatstat.sparse::positiveIndex' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::replacementIndex' by 'spatstat.sparse::replacementIndex' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::RelevantEmpty' by 'spatstat.sparse::RelevantEmpty' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::isRelevantZero' by 'spatstat.sparse::isRelevantZero' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::logicalIndex' by 'spatstat.sparse::logicalIndex' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::RelevantZero' by 'spatstat.sparse::RelevantZero' when loading 'spatstat.core'

#> Warning: replacing previous import 'spatstat.utils::strictIndexSequence' by 'spatstat.sparse::strictIndexSequence' when loading 'spatstat.core'