Convert gene IDs

Convert the gene IDs in a SingleCellExperiment object using Bioconductor org.*.eg.db data packages. Because annotation databases do not have a 1:1 relationship, this tool removes rows with no corresponding annotation in your desired annotation, and remove any duplicate annotations after conversion.

convert_gene_ids(inSCESet, in_symbol, out_symbol, database = "org.Hs.eg.db")

Arguments

inSCESet
Input SCtkExperiment object. Required

in_symbol
The input symbol type

out_symbol
The output symbol type

database
The org.*.eg.db database to use. The default is org.Hs.eg.db

Value

A SCtkExperiment with converted gene IDs.

Examples

if(requireNamespace("org.Mm.eg.db", quietly = TRUE)) {
  #convert mouse gene symbols to ensembl IDs
  library("org.Mm.eg.db")
  sample(rownames(GSE60361_subset_sce), 50)
  GSE60361_subset_symbol <- convert_gene_ids(inSCESet = GSE60361_subset_sce,
                                             in_symbol = "SYMBOL",
                                             out_symbol = "ENSEMBL",
                                             database = "org.Mm.eg.db")
  sample(rownames(GSE60361_subset_symbol), 50)
}

#> 

#> Loading required package: AnnotationDbi

#> 'select()' returned 1:many mapping between keys and columns

#> 1773 ENSEMBL are NA after conversion. Removing

#> 51 ENSEMBL are duplicated after conversion. Removing additional copies

#> 153 SYMBOL are duplicated after conversion. Removing additional copies

#> 19972 SYMBOL originally, 18153 ENSEMBLs

#>  [1] "ENSMUSG00000024781" "ENSMUSG00000014763" "ENSMUSG00000022477"
#>  [4] "ENSMUSG00000029603" "ENSMUSG00000052821" "ENSMUSG00000028309"
#>  [7] "ENSMUSG00000067578" "ENSMUSG00000018500" "ENSMUSG00000030353"
#> [10] "ENSMUSG00000026197" "ENSMUSG00000033632" "ENSMUSG00000095075"
#> [13] "ENSMUSG00000024673" "ENSMUSG00000020120" "ENSMUSG00000037112"
#> [16] "ENSMUSG00000063952" "ENSMUSG00000024050" "ENSMUSG00000055013"
#> [19] "ENSMUSG00000021681" "ENSMUSG00000018068" "ENSMUSG00000004415"
#> [22] "ENSMUSG00000079197" "ENSMUSG00000045288" "ENSMUSG00000040724"
#> [25] "ENSMUSG00000001445" "ENSMUSG00000003199" "ENSMUSG00000038644"
#> [28] "ENSMUSG00000020930" "ENSMUSG00000022914" "ENSMUSG00000074909"
#> [31] "ENSMUSG00000067851" "ENSMUSG00000033671" "ENSMUSG00000047497"
#> [34] "ENSMUSG00000033009" "ENSMUSG00000020674" "ENSMUSG00000039081"
#> [37] "ENSMUSG00000011463" "ENSMUSG00000030785" "ENSMUSG00000072944"
#> [40] "ENSMUSG00000056091" "ENSMUSG00000079620" "ENSMUSG00000074274"
#> [43] "ENSMUSG00000039196" "ENSMUSG00000008540" "ENSMUSG00000079557"
#> [46] "ENSMUSG00000046456" "ENSMUSG00000025665" "ENSMUSG00000114797"
#> [49] "ENSMUSG00000029797" "ENSMUSG00000034744"