Create a SCtkExperiment object

From a file of counts and a file of annotation information, create a SCtkExperiment object.

createSCE(
  assayFile = NULL,
  annotFile = NULL,
  featureFile = NULL,
  assayName = "counts",
  inputDataFrames = FALSE,
  createLogCounts = TRUE
)

Arguments

assayFile
The path to a text file that contains a header row of sample names, and rows of raw counts per gene for those samples.

annotFile
The path to a text file that contains columns of annotation information for each sample in the assayFile. This file should have the same number of rows as there are columns in the assayFile.

featureFile
The path to a text file that contains columns of annotation information for each gene in the count matrix. This file should have the same genes in the same order as assayFile. This is optional.

assayName
The name of the assay that you are uploading. The default is "counts".

inputDataFrames
If TRUE, assayFile and annotFile are read as data frames instead of file paths. The default is FALSE.

createLogCounts
If TRUE, create a log2(counts+1) normalized assay and include it in the object. The default is TRUE

Value

a SCtkExperiment object

Examples

data("mouseBrainSubsetSCE")
counts_mat <- assay(mouseBrainSubsetSCE, "counts")
sample_annot <- colData(mouseBrainSubsetSCE)
row_annot <- rowData(mouseBrainSubsetSCE)
newSCE <- createSCE(assayFile = counts_mat, annotFile = sample_annot,
                    featureFile = row_annot, assayName = "counts",
                    inputDataFrames = TRUE, createLogCounts = TRUE)