Helper function for differential expression analysis methods that accepts multiple ways of conditional subsetting and returns stable index format. Meanwhile it does all the input checkings.

Helper function for differential expression analysis methods that accepts multiple ways of conditional subsetting and returns stable index format. Meanwhile it does all the input checkings.

.formatDEAList(
  inSCE,
  useAssay,
  useReducedDim,
  index1 = NULL,
  index2 = NULL,
  class = NULL,
  classGroup1 = NULL,
  classGroup2 = NULL,
  groupName1,
  groupName2,
  analysisName,
  covariates = NULL,
  overwrite = FALSE
)

Arguments

inSCE

SingleCellExperiment inherited object. Required.

useAssay

character. A string specifying which assay to use. Required.

useReducedDim

character. A string specifying which reducedDim to use for DE analysis. Usually a pathway analysis result matrix. Set useAssay to NULL when using. Required.

index1

Any type of indices that can subset a SingleCellExperiment inherited object by cells. Specifies which cells are of interests. Default NULL.

index2

Any type of indices that can subset a SingleCellExperiment inherited object by cells. specifies the control group against those specified by index1. If NULL when using index specification, index1 cells will be compared with all other cells. Default NULL.

class

A vector/factor with ncol(inSCE) elements, or a character scalar that specifies a column name of colData(inSCE). Default NULL.

classGroup1

a vector specifying which "levels" given in class are of interests. Default NULL.

classGroup2

a vector specifying which "levels" given in class is the control group against those specified by classGroup1. If NULL when using annotation specification, classGroup1 cells will be compared with all other cells.

groupName1

A character scalar naming the group of interests. Required.

groupName2

A character scalar naming the control group. Required.

analysisName

A character scalar naming the DEG analysis. Required

covariates

A character vector of additional covariates used in linear regression methods such as Limma and DESeq2. Default NULL

overwrite

A logical scalar. Whether to overwrite result if exists. Default FALSE.

Value

A list object with part of formatted DE analysis information

Author

Yichen Wang