Plot results of reduced dimensions data.

Plot results of reduced dimensions data and colors the plots by the input vector.

.ggScatter(
  inSCE,
  sample = NULL,
  colorBy = NULL,
  groupBy = NULL,
  shape = NULL,
  reducedDimName,
  conditionClass = NULL,
  labelClusters = FALSE,
  xlab = NULL,
  ylab = NULL,
  axisSize = 10,
  axisLabelSize = 10,
  dim1 = NULL,
  dim2 = NULL,
  bin = NULL,
  binLabel = NULL,
  dotSize = 2,
  transparency = 1,
  colorScale = NULL,
  colorLow = "white",
  colorMid = "gray",
  colorHigh = "blue",
  defaultTheme = TRUE,
  title = NULL,
  titleSize = 15,
  legendTitle = NULL,
  legendTitleSize = 12,
  legendSize = 10,
  combinePlot = "none",
  plotLabels = NULL
)

Arguments

inSCE
Input SingleCellExperiment object with saved dimension reduction components or a variable with saved results. Required

sample
Character vector. Indicates which sample each cell belongs to.

colorBy
If provided, colors dots in the scatterplot based on value.

groupBy
If provided, facet wrap the scatterplot based on value.

shape
If provided, add shapes based on the value.

reducedDimName
Saved dimension reduction name in the SingleCellExperiment object. Required.

conditionClass
class of the annotation data used in colorBy. Options are NULL, "factor" or "numeric". If NULL, class will default to the original class. Default NULL.

labelClusters
Logical. Whether the cluster labels are plotted. Default FALSE.

xlab
Character vector. Label for x-axis. Default NULL.

ylab
Character vector. Label for y-axis. Default NULL.

axisSize
Size of x/y-axis ticks. Default 10.

axisLabelSize
Size of x/y-axis labels. Default 10.

dim1
1st dimension to be used for plotting. Can either be a string which specifies the name of the dimension to be plotted from reducedDims, or a numeric value which specifies the index of the dimension to be plotted. Default is NULL.

dim2
2nd dimension to be used for plotting. Can either be a string which specifies the name of the dimension to be plotted from reducedDims, or a numeric value which specifies the index of the dimension to be plotted. Default is NULL.

bin
Numeric vector. If single value, will divide the numeric values into the `bin` groups. If more than one value, will bin numeric values using values as a cut point.

binLabel
Character vector. Labels for the bins created by the `bin` parameter. Default NULL.

dotSize
Size of dots. Default 2.

transparency
Transparency of the dots, values will be 0-1. Default 1.

colorScale
Vector. Needs to be same length as the number of unique levels of `colorBy`. Will be used only if conditionClass = "factor" or "character". Default NULL.

colorLow
Character. A color available from `colors()`. The color will be used to signify the lowest values on the scale. Default 'white'. Will be used only if conditionClass = "numeric".

colorMid
Character. A color available from `colors()`. The color will be used to signify the midpoint on the scale. Default 'gray'. Will be used only if conditionClass = "numeric".

colorHigh
Character. A color available from `colors()`. The color will be used to signify the highest values on the scale. Default 'blue'. Will be used only if conditionClass = "numeric".

defaultTheme
Removes grid in plot and sets axis title size to 10 when TRUE. Default TRUE.

title
Title of plot. Default NULL.

titleSize
Size of title of plot. Default 15.

legendTitle
title of legend. Default NULL.

legendTitleSize
size of legend title. Default 12.

legendSize
size of legend. Default 10.

combinePlot
Must be either "all", "sample", or "none". "all" will combine all plots into a single .ggplot object, while "sample" will output a list of plots separated by sample. Default "none".

plotLabels
labels to each plot. If set to "default", will use the name of the samples as the labels. If set to "none", no label will be plotted.

Value

a ggplot of the reduced dimensions.