enrichR Given a list of genes this function runs the enrichR() to perform Gene enrichment

enrichR Given a list of genes this function runs the enrichR() to perform Gene enrichment

enrichRSCE(inSCE, glist, db = NULL)

Arguments

inSCE
Input SingleCellExperiment object.

glist
selected genes for enrichment analysis using enrichR(). Required

db
selected database name from the enrichR database list. if NULL then enrichR will be run on all the available databases on the enrichR database.

Value

enrichRSCE(): returns a data.frame of enrichment terms overlapping in the respective databases along with p-values, z-scores etc.,

Examples

if (FALSE) {
enrichRSCE(mouseBrainSubsetSCE, "Cmtm5", "GO_Cellular_Component_2017")
}