expData Store data items using tags to identify the type of data item stored. To be used as a replacement for assay<- setter function but with additional parameter to set a tag to a data item.

expData Store data items using tags to identify the type of data item stored. To be used as a replacement for assay<- setter function but with additional parameter to set a tag to a data item.

expData(inSCE, assayName, tag = NULL, altExp = FALSE) <- value

Arguments

inSCE
Input SingleCellExperiment object.

assayName
Specify the name of the input assay.

tag
Specify the tag to store against the input assay. Default is NULL, which will set the tag to "uncategorized".

altExp
A logical value indicating if the input assay is a altExp or a subset assay.

value
An input matrix-like value to store in the SCE object.