expTaggedData Returns a list of names of data items from the input SingleCellExperiment object based upon the input parameters.

expTaggedData Returns a list of names of data items from the input SingleCellExperiment object based upon the input parameters.

expTaggedData(inSCE, tags = NULL, redDims = FALSE)

Arguments

inSCE
Input SingleCellExperiment object.

tags
A character() value indicating if the data items should be returned separated by the specified tags. Default is NULL indicating that returned names of the data items are simply returned as a list with default tag as "uncategorized".

redDims
A logical value indicating if reducedDims should be returned as well separated with 'redDims' tag.

Value

A list of names of data items specfied by the other parameters.