Filter Genes and Samples from a Single Cell Object

Filter Genes and Samples from a Single Cell Object

filterSCData(
  inSCE,
  useAssay = "counts",
  deletesamples = NULL,
  removeNoExpress = TRUE,
  removeBottom = 0.5,
  minimumDetectGenes = 1700,
  filterSpike = TRUE
)

Arguments

inSCE
Input SingleCellExperiment object. Required

useAssay
Indicate which assay to use for filtering. Default is "counts"

deletesamples
List of samples to delete from the object.

removeNoExpress
Remove genes that have no expression across all samples. The default is true

removeBottom
Fraction of low expression genes to remove from the single cell object. This occurs after removeNoExpress. The default is 0.50.

minimumDetectGenes
Minimum number of genes with at least 1 count to include a sample in the single cell object. The default is 1700.

filterSpike
Apply filtering to Spike in controls (indicated by isSpike). The default is TRUE.

Value

The filtered single cell object.

Examples

if (FALSE) {
data("mouseBrainSubsetSCE")
mouseBrainSubsetSCE <- filterSCData(mouseBrainSubsetSCE,
                                    deletesamples="X1772063061_G11")
}