Find the marker gene set for each cluster With an input SingleCellExperiment object and specifying the clustering labels, this function iteratively call the differential expression analysis on each cluster against all the others.

Find the marker gene set for each cluster With an input SingleCellExperiment object and specifying the clustering labels, this function iteratively call the differential expression analysis on each cluster against all the others.

findMarkerDiffExp(
  inSCE,
  useAssay = "logcounts",
  method = c("wilcox", "MAST", "DESeq2", "Limma", "ANOVA"),
  cluster = "cluster",
  covariates = NULL,
  log2fcThreshold = 0.25,
  fdrThreshold = 0.05,
  minClustExprPerc = 0.6,
  maxCtrlExprPerc = 0.4,
  minMeanExpr = 0.5
)

Arguments

inSCE
SingleCellExperiment inherited object.

useAssay
character. A string specifying which assay to use for the MAST calculations. Default "logcounts".

method
A single character for specific differential expression analysis method. Choose from 'wilcox', 'MAST', 'DESeq2', 'Limma', and 'ANOVA'. Default "wilcox".

cluster
One single character to specify a column in colData(inSCE) for the clustering label. Alternatively, a vector or a factor is also acceptable. Default "cluster".

covariates
A character vector of additional covariates to use when building the model. All covariates must exist in names(colData(inSCE)). Not applicable when method is "MAST" method. Default NULL.

log2fcThreshold
Only out put DEGs with the absolute values of log2FC larger than this value. Default NULL

fdrThreshold
Only out put DEGs with FDR value smaller than this value. Default 1

minClustExprPerc
A numeric scalar. The minimum cutoff of the percentage of cells in the cluster of interests that expressed the marker gene. Default 0.7.

maxCtrlExprPerc
A numeric scalar. The maximum cutoff of the percentage of cells out of the cluster (control group) that expressed the marker gene. Default 0.4.

minMeanExpr
A numeric scalar. The minimum cutoff of the mean expression value of the marker in the cluster of interests. Default 1.

Value

The input SingleCellExperiment object with metadata(inSCE)$findMarker updated with a data.table of the up- regulated DEGs for each cluster.

Examples

data("mouseBrainSubsetSCE", package = "singleCellTK")
mouseBrainSubsetSCE <- findMarkerDiffExp(mouseBrainSubsetSCE,
                                         useAssay = "logcounts",
                                         cluster = "level1class")

#> Running with wilcox

#> Computing for cluster: microglia

#> Computing for cluster: oligodendrocytes