Generate manifest file for droplet and cell count data

Generate manifest file for droplet and cell count data

generateMeta(
  dropletSCE = NULL,
  cellSCE = NULL,
  samplename,
  dir,
  HTAN = TRUE,
  dataType = c("Droplet", "Cell", "Both")
)

Arguments

dropletSCE
A SingleCellExperiment object containing droplet count matrix data

cellSCE
A SingleCellExperiment object containing cell count matrix data

samplename
The sample name of the SingleCellExperiment objects

dir
The output directory of the SCTK QC pipeline.

HTAN
Whether generate manifest file with the format required by HTAN. Default is TRUE.

dataType
Type of the input data. It can be one of "Droplet", "Cell" or "Both".

Value

A SingleCellExperiment object which combines all objects in sceList. The colData is merged.

Examples

data(scExample, package = "singleCellTK")
generateMeta(sce, dir = ".", samplename = "Sample",
 HTAN = TRUE)

#> [[1]]
#>     SAMPLE      DATA_TYPE MATRIX_TYPE   DATA_CATEGORY CELL_MEDIAN_NUM_READS
#> Raw Sample Droplet Matrix  Raw Counts Gene Expression                  <NA>
#>     CELL_MEDIAN_NUM_GENES CELL_TOTAL
#> Raw                  <NA>        390
#>                                                  FILE_NAME
#> Raw ./Sample/FlatFile/Droplets/assays/Sample_counts.mtx.gz
#>                WORKFLOW_TYPE
#> Raw singleCellTK QC pipeline
#>                                                   WORKFLOW_PARAMETERS
#> Raw Ran perCellQC, EmptyDrops and barcodeRankDrops using singleCellTK
#>       WORKFLOW_VERSION HTAN_BIOSPECIMEN_ID HTAN_PARENT_ID HTAN_PARENT_FILE_ID
#> Raw singleCellTK:2.2.0              Sample                                   
#> 
#> [[2]]
#> NULL
#>