Get Top Table of a DEG analysis

Users have to run runDEAnalysis() first, any of the wrapped functions of this generic function. Users can set further filters on the result. A data.frame object, with variables of Gene, Log2_FC, Pvalue, and FDR, will be returned.

getDEGTopTable(
  inSCE,
  useResult,
  labelBy = NULL,
  onlyPos = FALSE,
  log2fcThreshold = 0.25,
  fdrThreshold = 0.05
)

Arguments

inSCE
SingleCellExperiment inherited object, with of the singleCellTK DEG method performed in advance.

useResult
character. A string specifying the analysisName used when running a differential expression analysis function.

labelBy
A single character for a column of rowData(inSCE) as where to search for the labeling text. Default NULL for the "rownames".

onlyPos
logical. Whether to only fetch DEG with positive log2_FC value. Default FALSE.

log2fcThreshold
numeric. Only fetch DEGs with the absolute values of log2FC larger than this value. Default 0.25.

fdrThreshold
numeric. Only fetch DEGs with FDR value smaller than this value. Default 0.05.

Value

A data.frame object of the top DEGs, with variables of Gene, Log2_FC, Pvalue, and FDR.

Examples

data("sceBatches")
sceBatches <- scaterlogNormCounts(sceBatches, "logcounts")
sce.w <- subsetSCECols(sceBatches, colData = "batch == 'w'")
sce.w <- runWilcox(sce.w, class = "cell_type", classGroup1 = "alpha",
                   groupName1 = "w.alpha", groupName2 = "w.beta",
                   analysisName = "w.aVSb")
getDEGTopTable(sce.w, "w.aVSb")

#>               Gene     Log2_FC       Pvalue          FDR
#> 74         ADCYAP1  -9.5222344 3.683325e-21 3.683325e-19
#> 57            HADH -10.5612840 6.226069e-20 3.113034e-18
#> 1              INS -12.7028733 2.368614e-19 4.737227e-18
#> 2              GCG   9.1517909 2.368614e-19 4.737227e-18
#> 29        INS-IGF2 -12.9324068 2.367422e-19 4.737227e-18
#> 16            IAPP -11.3726919 1.027650e-17 1.712751e-16
#> 88             TTR   4.3825672 3.348780e-17 4.783971e-16
#> 48              GC   7.1878882 4.658205e-15 5.822756e-14
#> 89            SCGN  -3.8122943 3.019451e-14 3.354946e-13
#> 34           MAP1B  -4.0106181 6.802842e-14 6.802842e-13
#> 9           TM4SF4   6.4072019 2.739759e-13 2.490690e-12
#> 86            RBP4  -5.5673377 4.757801e-11 3.964834e-10
#> 46          DPYSL2  -3.6028768 3.064074e-10 2.356980e-09
#> 56            PTEN  -3.3372373 3.758369e-10 2.684550e-09
#> 72            RGS4   5.4637919 9.524267e-10 6.349511e-09
#> 32          SEC22B  -2.7715616 2.049612e-09 1.281007e-08
#> 26       LOC647979  -2.9204542 2.744513e-09 1.614419e-08
#> 37           G6PC2  -4.5526717 4.065000e-09 2.258333e-08
#> 67         FAM18B1  -3.4609207 5.739256e-09 3.020661e-08
#> 22         SLC30A8  -3.0773276 1.216120e-08 6.080598e-08
#> 53           ETNK1  -3.4238773 8.381062e-08 3.990982e-07
#> 28        C10orf46  -4.1172723 1.264781e-07 5.749006e-07
#> 71            TLK1  -2.7498540 1.541148e-07 6.700642e-07
#> 61            PDK4   4.5360064 2.535638e-07 1.056516e-06
#> 18    NM-001166106  -1.7095593 4.756783e-07 1.902713e-06
#> 100        C8orf83  -3.7073784 6.952554e-07 2.674059e-06
#> 39            RIN2  -2.6455051 1.481920e-06 5.488594e-06
#> 80            MLL5  -2.3693014 6.750971e-06 2.411061e-05
#> 98  RPL17-C18ORF32  -1.7895368 8.133631e-06 2.804700e-05
#> 79          TMEM85  -2.5089275 1.018119e-05 3.393730e-05
#> 44           SRSF6  -2.0881730 1.189004e-05 3.835497e-05
#> 66          C9orf5  -2.9296536 3.752854e-05 1.172767e-04
#> 47            TSIX  -3.9086902 4.398954e-05 1.333016e-04
#> 70          TUBA1A  -1.8223863 4.544070e-05 1.336491e-04
#> 59           SC5DL  -3.2686407 4.937434e-05 1.410695e-04
#> 77          ZNF252  -3.1415988 6.620272e-05 1.838964e-04
#> 51         C2orf28  -2.3401725 1.039625e-04 2.809797e-04
#> 68        SERPINA1   2.9853497 2.218207e-04 5.837387e-04
#> 40            VCAN  -1.6914180 2.586650e-04 6.632435e-04
#> 33             CHP  -2.0821137 2.684848e-04 6.712119e-04
#> 20            EGR1  -2.0661563 3.793268e-04 9.251873e-04
#> 43            MLL3  -1.7807672 4.587914e-04 1.077738e-03
#> 96        C11orf10  -1.2673860 4.634275e-04 1.077738e-03
#> 78        KCNQ1OT1   2.6699417 6.608041e-04 1.501827e-03
#> 69             MLL  -1.8153392 1.348475e-03 2.996612e-03
#> 63       LOC729013  -2.5631018 1.561273e-03 3.394072e-03
#> 64         C6orf62  -1.4303264 1.776495e-03 3.779777e-03
#> 99           TTC35  -2.2703597 1.915506e-03 3.990638e-03
#> 52        C14orf43  -2.6597468 2.774656e-03 5.662562e-03
#> 97           SNX22  -1.1238276 3.454043e-03 6.908086e-03
#> 81            PKM2  -1.1729284 6.785609e-03 1.330512e-02
#> 75            ROD1  -1.8832646 7.790889e-03 1.498248e-02
#> 85        C10orf58  -1.6580742 8.496465e-03 1.603107e-02
#> 83        FLJ43390  -2.1821390 1.284192e-02 2.378134e-02
#> 11            PENK   0.9654197 1.754329e-02 3.189689e-02
#> 73             FOS  -1.3535129 1.823620e-02 3.256465e-02
#> 87        KIAA0494  -1.8671111 2.676761e-02 4.696072e-02