Run t-SNE dimensionality reduction method on a SingleCellExperiment Object

Run t-SNE dimensionality reduction method on a SingleCellExperiment Object

getTSNE(
  inSCE,
  useAssay = "logcounts",
  useAltExp = NULL,
  useReducedDim = NULL,
  reducedDimName = "TSNE",
  nIterations = 1000,
  perplexity = 30,
  run_pca = TRUE,
  ntop = NULL,
  seed = NULL
)

Arguments

inSCE
Input SingleCellExperiment object.

useAssay
Assay to use for tSNE computation. If useAltExp is specified, useAssay has to exist in assays(altExp(inSCE, useAltExp)). Default "logcounts"

useAltExp
The subset to use for tSNE computation, usually for the selected.variable features. Default NULL.

useReducedDim
The low dimension representation to use for UMAP computation. Default NULL.

reducedDimName
a name to store the results of the dimension reductions. Default "TSNE".

nIterations
maximum iterations. Default 1000.

perplexity
perplexity parameter. Default 30.

run_pca
run tSNE on PCA components? Default TRUE.

ntop
Number of top features to use as a further variable feature selection. Default NULL.

seed
Random seed for reproducibility of tSNE results. Default NULL will use global seed in use by the R environment.

Value

A SingleCellExperiment object with tSNE computation updated in reducedDim(inSCE, reducedDimName).

Examples

data("mouseBrainSubsetSCE")
#add a CPM assay
assay(mouseBrainSubsetSCE, "cpm") <- apply(
  assay(mouseBrainSubsetSCE, "counts"), 2, function(x) {
    x / (sum(x) / 1000000)
  })
mouseBrainSubsetSCE <- getTSNE(mouseBrainSubsetSCE, useAssay = "cpm",
                               reducedDimName = "TSNE_cpm",
                               perplexity = NULL)