Retrieve example datasets

Retrieves published example datasets stored in SingleCellExperiment using the scRNAseq and TENxPBMCData packages. See 'Details' for a list of available datasets.

importExampleData(dataset, class = c("Matrix", "matrix"), delayedArray = FALSE)

Arguments

dataset
Character. Name of the dataset to retrieve.

class
Character. The class of the expression matrix stored in the SCE object. Can be one of "Matrix" or "matrix". "Matrix" will store the data as a sparse matrix from package Matrix while "matrix" will store the data in a standard matrix. Default "Matrix".

delayedArray
Boolean. Whether to read the expression matrix as DelayedArray object or not. Default FALSE.

Value

The specified SingleCellExperiment object.

Details

See the list below for the available datasets and their descriptions.

"fluidigm_pollen"

Retrieved with ReprocessedFluidigmData. Returns a dataset of 65 human neural cells from Pollen et al. (2014), each sequenced at high and low coverage (SRA accession SRP041736).

"allen_tasic"

Retrieved with ReprocessedAllenData. Returns a dataset of 379 mouse brain cells from Tasic et al. (2016).

"pbmc3k"

Retrieved with TENxPBMCData. 2,700 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc4k"

Retrieved with TENxPBMCData. 4,340 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc6k"

Retrieved with TENxPBMCData. 5,419 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc8k"

Retrieved with TENxPBMCData. 8,381 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc33k"

Retrieved with TENxPBMCData. 33,148 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc68k"

Retrieved with TENxPBMCData. 68,579 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

Author

Joshua D. Campbell, David Jenkins

Examples

sce <- importExampleData("pbmc3k")

#> snapshotDate(): 2020-10-27

#> see ?TENxPBMCData and browseVignettes('TENxPBMCData') for documentation

#> loading from cache