Get runDropletQC .html report

A function to generate .html Rmarkdown report containing the visualizations of the runDropletQC function output

reportDropletQC(
  inSCE,
  output_file = NULL,
  output_dir = NULL,
  subTitle = NULL,
  studyDesign = NULL
)

Arguments

inSCE

A SingleCellExperiment object containing the full droplet count matrix with the output from runDropletQC function

output_file

name of the generated file. If NULL/default then the output file name will be based on the name of the Rmarkdown template

output_dir

name of the output directory to save the rendered file. If NULL/default the file is stored to the current working directory

subTitle

subtitle of the QC HTML report. Default is NULL.

studyDesign

description of the data set and experiment design. It would be shown at the top of QC HTML report. Default is NULL.

Value

.html file

Examples

data(scExample, package = "singleCellTK")
if (FALSE) {
sce <- runDropletQC(sce)
reportDropletQC(inSCE = sce)
}