Get .html report of the output of the selected QC algorithm

A function to generate .html Rmarkdown report for the specified QC algorithm output

reportQCTool(
  inSCE,
  algorithm = c("BarcodeRankDrops", "EmptyDrops", "QCMetrics", "Scrublet",
    "DoubletCells", "Cxds", "Bcds", "CxdsBcdsHybrid", "DoubletFinder", "DecontX"),
  output_file = NULL,
  output_dir = NULL
)

Arguments

inSCE
A SingleCellExperiment object containing the count matrix (full droplets or filtered matrix, depends on the selected QC algorithm) with the output from at least one of these functions: runQCMetrics, runScrublet, runDoubletCells, runCxds, runBcds, runCxdsBcdsHybrid, runDecontX, runBarcodeRankDrops, runEmptyDrops

algorithm
Character. Specifies which QC algorithm report to generate. Available options are "BarcodeRankDrops", "EmptyDrops", "QCMetrics", "Scrublet", "DoubletCells", "Cxds", "Bcds", "CxdsBcdsHybrid", "DoubletFinder" and "DecontX".

output_file
name of the generated file. If NULL/default then the output file name will be based on the name of the selected QC algorithm name .

output_dir
name of the output directory to save the rendered file. If NULL/default the file is stored to the current working directory

Value

.html file

Examples

# \donttest{
data(scExample, package = "singleCellTK")
sce <- sce[, colData(sce)$type != 'EmptyDroplet']
sce <- runDecontX(sce)

#> Wed Aug 26 12:43:55 2020 ... Running 'DecontX'

#> --------------------------------------------------

#> Starting DecontX

#> --------------------------------------------------

#> Wed Aug 26 12:43:55 2020 .. Analyzing all cells

#> Wed Aug 26 12:43:55 2020 .... Generating UMAP and estimating cell types

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by ‘spam’

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by ‘spam’

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by ‘spam’

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by ‘spam’

#> Wed Aug 26 12:43:57 2020 .... Estimating contamination

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 10 | converge: 0.0278

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 20 | converge: 0.01058

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 30 | converge: 0.00584

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 40 | converge: 0.004697

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 50 | converge: 0.003356

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 60 | converge: 0.002634

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 70 | converge: 0.001941

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 80 | converge: 0.001413

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 90 | converge: 0.001284

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 100 | converge: 0.00111

#> Wed Aug 26 12:43:57 2020 ...... Completed iteration: 104 | converge: 0.0009827

#> Wed Aug 26 12:43:58 2020 .. Finalizing decontaminated matrix

#> --------------------------------------------------

#> Completed DecontX. Total time: 2.954186 secs

#> --------------------------------------------------

reportQCTool(inSCE = sce, algorithm = "DecontX")

#> 
#> 
#> processing file: DecontX.Rmd

#>   |                                                                              |                                                                      |   0%  |                                                                              |........                                                              |  11%
#>    inline R code fragments
#> 
#>   |                                                                              |................                                                      |  22%
#> label: decontX-lib (with options) 
#> List of 2
#>  $ message: symbol F
#>  $ include: logi FALSE
#> 

#> Loading required package: umap

#> Loading required package: ggplot2

#>   |                                                                              |.......................                                               |  33%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................                                       |  44%
#> label: decontX-input (with options) 
#> List of 1
#>  $ warning: logi FALSE
#> 
#>   |                                                                              |.......................................                               |  56%
#>   ordinary text without R code
#> 
#>   |                                                                              |...............................................                       |  67%
#> label: DecontX (with options) 
#> List of 4
#>  $ results  : chr "asis"
#>  $ fig.align: chr "center"
#>  $ warning  : logi FALSE
#>  $ message  : logi FALSE
#> 

#> Quitting from lines 24-41 (DecontX.Rmd) 

#> Error in `rownames<-`(`*tmp*`, value = c("pbmc_4k_GACCAATTCCTAGAAC-1", "pbmc_4k_AAATGCCGTTTCGCTC-1", "pbmc_4k_ATAACGCAGTGTCCAT-1", "pbmc_4k_ATAAGAGTCCGATATG-1", "pbmc_4k_CATGACATCGGATGTT-1", "pbmc_4k_GGCGACTGTTCTCATT-1", "pbmc_4k_ATGAGGGAGAGTCGGT-1", "pbmc_4k_AGCTCTCCAGCTTCGG-1", "pbmc_4k_GGAATAAGTTTGACTG-1", "pbmc_4k_TCGCGTTGTGTGACCC-1", "pbmc_4k_GACTGCGAGGCATGTG-1", "pbmc_4k_CGTTGGGCATCCGGGT-1", "pbmc_4k_TACGGATAGAAACGAG-1", "pbmc_4k_AACGTTGAGGTCATCT-1", "pbmc_4k_CAGAATCTCTCTGAGA-1", "pbmc_4k_CGGACTGAGATGTGTA-1", "pbmc_4k_ATTACTCAGTGATCGG-1", "pbmc_4k_CTAACTTTCAACACGT-1", "pbmc_4k_TGAAAGACATCTCCCA-1", "pbmc_4k_CAAGAAATCCAGAAGG-1", "pbmc_4k_CAGAGAGAGCTAAACA-1", "pbmc_4k_CAAGATCGTAAATGTG-1", "pbmc_4k_CTACACCTCTTGCAAG-1", "pbmc_4k_CCACTACCACGGACAA-1", "pbmc_4k_CATTCGCGTACTCAAC-1", "pbmc_4k_CACACAAGTACATGTC-1", "pbmc_4k_ATAACGCTCCTGTAGA-1", "pbmc_4k_GTAGGCCGTACGCTGC-1", "pbmc_4k_GCTCTGTGTTAAGAAC-1", "pbmc_4k_GTTCGGGGTTCACCTC-1", "pbmc_4k_GGTGAAGAGTCTTGCA-1", "pbmc_4k_ACGTCAAGTAAGAGAG-1", "pbmc_4k_AAACGGGAGGCTAGCA-1", "pbmc_4k_TGGCGCAGTGGCAAAC-1", "pbmc_4k_TGTCCCAGTTCCACTC-1", "pbmc_4k_TAGACCATCAGATAAG-1", "pbmc_4k_TGCTACCGTACACCGC-1", "pbmc_4k_CTCGAAACAGTGAGTG-1", "pbmc_4k_TGAGCCGCAAGCCCAC-1", "pbmc_4k_AGCGTCGGTGTCAATC-1", "pbmc_4k_GAAATGATCGTCCAGG-1", "pbmc_4k_TGACAACTCCCTCTTT-1", "pbmc_4k_TCGCGTTTCATGGTCA-1", "pbmc_4k_CGACCTTTCATGTGGT-1", "pbmc_4k_AGGGAGTCACGGTGTC-1", "pbmc_4k_AGCGGTCAGCTCCCAG-1", "pbmc_4k_TTGCGTCCATCCTTGC-1", "pbmc_4k_GTAACTGCAGTATAAG-1", "pbmc_4k_TGAGGGAAGTACCGGA-1", "pbmc_4k_AAAGTAGAGCGACGTA-1", "pbmc_4k_CGTTAGAGTAATAGCA-1", "pbmc_4k_CACACAACAGTCTTCC-1", "pbmc_4k_AAGGTTCGTTCACGGC-1", "pbmc_4k_CGGCTAGTCTACCTGC-1", "pbmc_4k_GTCATTTAGAGACGAA-1", "pbmc_4k_GGGATGAAGCAGACTG-1", "pbmc_4k_CATATGGAGCTCCTTC-1", "pbmc_4k_TAGACCACACATAACC-1", "pbmc_4k_GGCTCGAAGGCTCATT-1", "pbmc_4k_AGCGGTCAGATCACGG-1", "pbmc_4k_TACGGTAGTCGAATCT-1", "pbmc_4k_ACTGTCCAGCCACGTC-1", "pbmc_4k_TTGGCAACATTAGGCT-1", "pbmc_4k_CTTAACTTCTCGAGTA-1", "pbmc_4k_ATAGACCAGTGGAGAA-1", "pbmc_4k_CGCGGTATCGCACTCT-1", "pbmc_4k_GCGAGAAAGAATGTGT-1", "pbmc_4k_AGCGTCGCACGGTGTC-1", "pbmc_4k_AACACGTGTACCGTAT-1", "pbmc_4k_ACGGCCAGTTAAAGAC-1", "pbmc_4k_CGACTTCCAAACGCGA-1", "pbmc_4k_TACGGATGTAGGGTAC-1", "pbmc_4k_TGCTGCTGTTTGTGTG-1", "pbmc_4k_CGATGTAGTAGAGCTG-1", "pbmc_4k_CTCTAATGTTGGACCC-1", "pbmc_4k_AGTCTTTCAGCTCGAC-1", "pbmc_4k_CGGACTGTCAGTTCGA-1", "pbmc_4k_TGTTCCGCAGCATGAG-1", "pbmc_4k_CAAGGCCTCTCGCTTG-1", "pbmc_4k_AGCCTAATCATTCACT-1", "pbmc_4k_CATGGCGGTCATGCAT-1", "pbmc_4k_CATTCGCCATATACCG-1", "pbmc_4k_TGAGGGAAGAGTACCG-1", "pbmc_4k_ATTACTCCACAAGTAA-1", "pbmc_4k_AAGGCAGCAAATCCGT-1", "pbmc_4k_TCTGGAAAGAGTACAT-1", "pbmc_4k_TCTATTGCAAGCGAGT-1", "pbmc_4k_TTAGGACAGCGTAATA-1", "pbmc_4k_CAGAGAGCAGTAGAGC-1", "pbmc_4k_GCGCCAAGTCATGCAT-1", "pbmc_4k_CGTTGGGAGGTGCTAG-1", "pbmc_4k_CAACCAAGTAAGAGGA-1", "pbmc_4k_GAACGGATCCCGGATG-1", "pbmc_4k_GACAGAGAGTACGTAA-1", "pbmc_4k_CCTTCGATCAGGCGAA-1", "pbmc_4k_GTATTCTAGGATGGAA-1", "pbmc_4k_TGAGCATAGCGTCAAG-1", "pbmc_4k_GCAGTTAGTGGCCCTA-1", "pbmc_4k_TGCCCATGTTATTCTC-1", "pbmc_4k_CCGTACTTCGATAGAA-1", "pbmc_4k_CTAACTTTCTAACCGA-1", "pbmc_4k_ATCTACTTCTGAGTGT-1", "pbmc_4k_TCCACACAGTCCGTAT-1", "pbmc_4k_AACGTTGAGCGATATA-1", "pbmc_4k_CTAAGACAGGAACTGC-1", "pbmc_4k_GTTCGGGCAGTAGAGC-1", "pbmc_4k_AGCATACCAAGCCATT-1", "pbmc_4k_ACGCCAGGTGAACCTT-1", "pbmc_4k_ACGAGCCTCGGAGCAA-1", "pbmc_4k_GTCACGGTCAGTCAGT-1", "pbmc_4k_CTGATCCGTTCGCGAC-1", "pbmc_4k_GTGTGCGGTCCTAGCG-1", "pbmc_4k_CGGCTAGCAGGAACGT-1", "pbmc_4k_AGGCCACAGTGTCCCG-1", "pbmc_4k_AGCAGCCAGATGTCGG-1", "pbmc_4k_TTCTCCTAGCACCGCT-1", "pbmc_4k_CACACCTAGTGAACGC-1", "pbmc_4k_CGTTGGGAGAGCTATA-1", "pbmc_4k_TTAGGCAAGCCGCCTA-1", "pbmc_4k_CCTCTGACATTACCTT-1", "pbmc_4k_TATTACCAGTGAACAT-1", "pbmc_4k_ACTGATGCACCTGGTG-1", "pbmc_4k_CTCACACCAGTTCATG-1", "pbmc_4k_CGCGGTAAGGGAACGG-1", "pbmc_4k_GACCTGGAGCCCAACC-1", "pbmc_4k_GTTCATTCATCTCCCA-1", "pbmc_4k_AAACGGGGTTTGCATG-1", "pbmc_4k_GCGCAACCAGGTGGAT-1", "pbmc_4k_ACTGTCCCATGTCGAT-1", "pbmc_4k_AGGTCATCAGCTTCGG-1", "pbmc_4k_GATCAGTCACAAGTAA-1", "pbmc_4k_TGTGGTATCACGGTTA-1", "pbmc_4k_CCTACACAGGTGACCA-1", "pbmc_4k_ACGTCAAGTTCAGCGC-1", "pbmc_4k_CTCATTAAGCCAACAG-1", "pbmc_4k_ACGATACAGCCCAGCT-1", "pbmc_4k_CCTACCATCTGTTGAG-1", "pbmc_4k_CCGTTCATCGTAGATC-1", "pbmc_4k_TCGTACCAGCTAAACA-1", "pbmc_4k_TTCGAAGCAGGGCATA-1", "pbmc_4k_GTCAAGTCATCCAACA-1", "pbmc_4k_CAAGTTGTCACGCATA-1", "pbmc_4k_TTCGGTCAGCGTTCCG-1", "pbmc_4k_GTCTTCGGTTCCACTC-1", "pbmc_4k_GCTGGGTCAGTATGCT-1", "pbmc_4k_CCGGGATCAGAGCCAA-1", "pbmc_4k_GAGCAGATCGCCATAA-1", "pbmc_4k_TTAACTCCACTTACGA-1", "pbmc_4k_TGGGAAGCACAGAGGT-1", "pbmc_4k_GACGGCTGTTGGTAAA-1", "pbmc_4k_CGCTATCCAGGTGCCT-1", "pbmc_4k_AGTGGGAGTGCCTGTG-1", "pbmc_4k_GCTGCGACATAACCTG-1", "pbmc_4k_GGATTACAGGCTATCT-1", "pbmc_4k_GGCAATTGTCGCTTCT-1", "pbmc_4k_CGTTCTGAGGCTCAGA-1", "pbmc_4k_TTAACTCAGCTAACAA-1", "pbmc_4k_CATCAAGTCACGACTA-1", "pbmc_4k_CAAGGCCGTCGCTTTC-1", "pbmc_4k_GGACAAGTCTCAAGTG-1", "pbmc_4k_GTGCTTCGTTGTCTTT-1", "pbmc_4k_CACTCCATCTGAAAGA-1", "pbmc_4k_TTAGTTCCATGTAAGA-1", "pbmc_4k_AGCTCCTGTATAAACG-1", "pbmc_4k_CCACCTAAGTGGAGAA-1", "pbmc_4k_TAAACCGGTAGCAAAT-1", "pbmc_4k_AGCTCTCGTTGATTGC-1", "pbmc_4k_TCTCATAAGATAGGAG-1", "pbmc_4k_TCACAAGTCCGTCAAA-1", "pbmc_4k_GTGCTTCAGGCTAGAC-1", "pbmc_4k_GGAAAGCTCGAATGCT-1", "pbmc_4k_AAGCCGCCAATCTGCA-1", "pbmc_4k_GTTAAGCAGTGAAGAG-1", "pbmc_4k_AAGGCAGGTAGTGAAT-1", "pbmc_4k_CGCGTTTGTAGATTAG-1", "pbmc_4k_GATCGTAGTCTAGGTT-1", "pbmc_4k_CGGACTGAGCTGCAAG-1", "pbmc_4k_CCACTACTCATGTAGC-1", "pbmc_4k_AAGGTTCGTGCCTGCA-1", "pbmc_4k_CACACTCCAGACAAGC-1", "pbmc_4k_ATCTGCCAGACAAAGG-1", "pbmc_4k_CGAATGTCATCACCCT-1", "pbmc_4k_ATCTACTGTTTAGCTG-1", "pbmc_4k_TCAGCAACAATGGAGC-1", "pbmc_4k_ACGATGTAGTCCAGGA-1", "pbmc_4k_AGAGTGGTCGGTCCGA-1", "pbmc_4k_GGTGCGTAGTACGATA-1", "pbmc_4k_CTCCTAGTCGTACCGG-1", "pbmc_4k_CGATGTATCGTACCGG-1", "pbmc_4k_GCTTGAAGTCGGATCC-1", "pbmc_4k_CGGACACTCGGTCCGA-1", "pbmc_4k_TGGGAAGAGTGGTCCC-1", "pbmc_4k_CTCCTAGAGCAGACTG-1", "pbmc_4k_ATCGAGTGTACAGCAG-1", "pbmc_4k_TGACTAGAGTACGCGA-1")): attempt to set 'rownames' on an object with no dimensions

#> 

# }