Generates an HTML report for Seurat Results (including Clustering & Marker Selection) and returns the SCE object with the results computed and stored inside the object.

Generates an HTML report for Seurat Results (including Clustering & Marker Selection) and returns the SCE object with the results computed and stored inside the object.

reportSeuratResults(
  inSCE,
  biological.group = NULL,
  phenotype.groups = NULL,
  selected.markers = NULL,
  clustering.resolution = 0.8,
  pc.count = 50,
  plotTSNE = TRUE,
  plotUMAP = TRUE,
  runClustering = TRUE,
  runMSClusters = TRUE,
  runMSBioGroup = TRUE,
  numTopFeatures = 10,
  outputFile = NULL,
  outputPath = NULL,
  subtitle = NULL,
  authors = NULL,
  showSession = FALSE,
  pdf = FALSE,
  forceRun = TRUE
)

Arguments

inSCE

Input SingleCellExperiment object previously passed through reportSeuratRun().

biological.group

A character value that specifies the name of the colData() column to use as the main biological group in the Seurat report for marker selection and grouping.

phenotype.groups

A character vector that specifies the names of the colData() columns to use for differential expression in addition to the biological.group parameter.

selected.markers

A character vector containing the user-specified gene symbols or feature names of marker genes that be used to generate gene plots in addition to the gene markers computed from differential expression.

clustering.resolution

A numeric value indicating the user-specified final resolution to use with clustering. Default is 0.8.

pc.count

A numeric value indicating the number of principal components to use in the analysis workflow. Default is 50.

plotTSNE

A logical value indicating if TSNE plots should be visualized in the clustering section of the report. Default is TRUE.

plotUMAP

A logical value indicating if UMAP plots should be visualized in the clustering section of the report. Default is TRUE.

runClustering

A logical value indicating if Clustering should be run or not in the report. Default is TRUE. If FALSE, parameters plotTSNE and plotUMAP are also set to FALSE.

runMSClusters

A logical value indicating if the marker selection section for identifying marker genes between clusters should be run and visualized in the report. Default TRUE.

runMSBioGroup

A logical value indicating if the marker selection section for identifying marker genes between the biological.group parameter should be run and visualized in the report. Default TRUE.

numTopFeatures

A numeric value indicating the number of top features to visualize in each group. Default 10.

outputFile

Specify the name of the generated output HTML file. If NULL then the output file name will be based on the name of the Rmarkdown template. Default NULL.

outputPath

Specify the name of the output directory to save the rendered HTML file. If NULL the file is stored to the current working directory. Default NULL.

subtitle

A character value specifying the subtitle to use in the report. Default NULL.

authors

A character value specifying the names of the authors to use in the report. Default NULL.

showSession

A logical value indicating if session information should be displayed or not. Default is FALSE.

pdf

A logical value indicating if a pdf should also be generated for each figure in the report. Default is FALSE.

forceRun

A logical value indicating if all computations previously computed should be re-calculated regardless if these computations are available in the input object. Default is TRUE.

Value

A SingleCellExperiment object with computations stored.