Perform comprehensive single cell QC

A wrapper function to run several QC algorithms on a SingleCellExperiment object containing cells after empty droplets have been removed.

runCellQC(
  inSCE,
  algorithms = c("QCMetrics", "doubletCells", "cxds", "bcds", "cxds_bcds_hybrid",
    "scrublet", "doubletFinder", "decontX"),
  sample = NULL,
  geneSetList = NULL,
  geneSetListLocation = "rownames",
  geneSetCollection = NULL,
  useAssay = "counts",
  seed = 12345,
  paramsList = NULL
)

Arguments

inSCE
A SingleCellExperiment object.

algorithms
Character vector. Specify which QC algorithms to run. Available options are "QCMetrics", "scrublet", "doubletCells", "cxds", "bcds", "cxds_bcds_hybrid", and "decontX".

sample
Character vector. Indicates which sample each cell belongs to. Algorithms will be run on cells from each sample separately.

geneSetList
See runPerCellQC. Default NULL.

geneSetListLocation
See runPerCellQC. Default NULL.

geneSetCollection
See runPerCellQC. Default NULL.

useAssay
A string specifying which assay contains the count matrix for cells.

seed
Seed for the random number generator. Default 12345.

paramsList
A list containing parameters for QC functions. Default NULL.

Value

SingleCellExperiment object containing the outputs of the specified algorithms in the colData of inSCE.

Examples

if (FALSE) {
data(sce_chcl, package = "scds")
sce <- runCellQC(sce_chcl)
}