Detecting contamination with DecontX.

A wrapper function for decontX. Identify potential contamination from experimental factors such as ambient RNA.

runDecontX(
  inSCE,
  sample = NULL,
  useAssay = "counts",
  z = NULL,
  maxIter = 500,
  delta = c(10, 10),
  estimateDelta = TRUE,
  convergence = 0.001,
  iterLogLik = 10,
  varGenes = 5000,
  dbscanEps = 1,
  seed = 12345,
  logfile = NULL,
  verbose = TRUE
)

Arguments

inSCE
A SingleCellExperiment object.

sample
A single character specifying a name that can be found in colData(inSCE) to directly use the cell annotation; or a character vector with as many elements as cells to indicates which sample each cell belongs to. Default NULL. decontX will be run on cells from each sample separately.

useAssay
A string specifying which assay in the SCE to use. Default 'counts'.

z
Numeric or character vector. Cell cluster labels. If NULL, PCA will be used to reduce the dimensionality of the dataset initially, 'umap' from the 'uwot' package will be used to further reduce the dataset to 2 dimenions and the 'dbscan' function from the 'dbscan' package will be used to identify clusters of broad cell types. Default NULL.

maxIter
Integer. Maximum iterations of the EM algorithm. Default 500.

delta
Numeric Vector of length 2. Concentration parameters for the Dirichlet prior for the contamination in each cell. The first element is the prior for the native counts while the second element is the prior for the contamination counts. These essentially act as pseudocounts for the native and contamination in each cell. If estimateDelta = TRUE, this is only used to produce a random sample of proportions for an initial value of contamination in each cell. Then fit_dirichlet is used to update delta in each iteration. If estimateDelta = FALSE, then delta is fixed with these values for the entire inference procedure. Fixing delta and setting a high number in the second element will force decontX to be more aggressive and estimate higher levels of contamination at the expense of potentially removing native expression. Default c(10, 10).

estimateDelta
Boolean. Whether to update delta at each iteration.

convergence
Numeric. The EM algorithm will be stopped if the maximum difference in the contamination estimates between the previous and current iterations is less than this. Default 0.001.

iterLogLik
Integer. Calculate log likelihood every iterLogLik iteration. Default 10.

varGenes
Integer. The number of variable genes to use in dimensionality reduction before clustering. Variability is calcualted using modelGeneVar function from the 'scran' package. Used only when z is not provided. Default 5000.

dbscanEps
Numeric. The clustering resolution parameter used in 'dbscan' to estimate broad cell clusters. Used only when z is not provided. Default 1.

seed
Integer. Passed to with_seed. For reproducibility, a default value of 12345 is used. If NULL, no calls to with_seed are made.

logfile
Character. Messages will be redirected to a file named `logfile`. If NULL, messages will be printed to stdout. Default NULL.

verbose
Logical. Whether to print log messages. Default TRUE.

Value

A SingleCellExperiment object with 'decontX_Contamination' and 'decontX_Clusters' added to the colData slot. Additionally, the decontaminated counts will be added as an assay called 'decontXCounts'.

Examples

data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runDecontX(sce[,sample(ncol(sce),20)])

#> Fri Oct 01 18:22:40 2021 ... Running 'DecontX'

#> --------------------------------------------------

#> Starting DecontX

#> --------------------------------------------------

#> Fri Oct 01 18:22:40 2021 .. Analyzing all cells

#> Fri Oct 01 18:22:40 2021 .... Generating UMAP and estimating cell types

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Fri Oct 01 18:22:42 2021 .... Estimating contamination

#> Fri Oct 01 18:22:42 2021 ...... Completed iteration: 10 | converge: 0.02037

#> Fri Oct 01 18:22:42 2021 ...... Completed iteration: 20 | converge: 0.007198

#> Fri Oct 01 18:22:42 2021 ...... Completed iteration: 30 | converge: 0.002494

#> Fri Oct 01 18:22:42 2021 ...... Completed iteration: 40 | converge: 0.001422

#> Fri Oct 01 18:22:42 2021 ...... Completed iteration: 50 | converge: 0.001155

#> Fri Oct 01 18:22:42 2021 ...... Completed iteration: 60 | converge: 0.0009939

#> Fri Oct 01 18:22:43 2021 .. Finalizing decontaminated matrix

#> --------------------------------------------------

#> Completed DecontX. Total time: 2.958157 secs

#> --------------------------------------------------