Wrapper function to run all of the feature selection methods integrated within the singleCellTK package including three methods from Seurat (`vst`, `mean.var.plot` or `dispersion`) and the Scran `modelGeneVar` method.

Wrapper function to run all of the feature selection methods integrated within the singleCellTK package including three methods from Seurat (`vst`, `mean.var.plot` or `dispersion`) and the Scran `modelGeneVar` method.

runFeatureSelection(
  inSCE,
  useAssay,
  hvgMethod = c("vst", "mean.var.plot", "dispersion", "modelGeneVar")
)

Arguments

inSCE

Input SingleCellExperiment object.

useAssay

Specify the name of the assay that should be used. A normalized assay is recommended for use with this function.

hvgMethod

Specify the method to use for variable gene selection. Options include "vst", "mean.var.plot" or "dispersion" from Seurat and "modelGeneVar" from Scran.

Value

A SingleCellExperiment object that contains the computed statistics in the rowData slot of the output object. This function does not return the names of the variable features but only computes the statistics that are stored in the rowData slot of the. To get the names of the variable features getTopHVG function should be used after computing these statistics.

Examples

data("mouseBrainSubsetSCE", package = "singleCellTK")
mouseBrainSubsetSCE <- runFeatureSelection(mouseBrainSubsetSCE,
                                           "logcounts",
                                           "modelGeneVar")
#> Warning: collapsing to unique 'x' values