Apply LIGER batch effect correction method to SingleCellExperiment object

LIGER relies on integrative non-negative matrix factorization to identify shared and dataset-specific factors.

runLIGER(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  reducedDimName = "LIGER",
  nComponents = 20L,
  lambda = 5,
  resolution = 1
)

Arguments

inSCE
SingleCellExperiment inherited object. Required.

useAssay
A single character indicating the name of the assay requiring batch correction. Default "logcounts".

batch
A single character indicating a field in colData that annotates the batches. Default "batch".

reducedDimName
A single character. The name for the corrected low-dimensional representation. Will be saved to reducedDim(inSCE). Default "LIGER".

nComponents
An integer. The number of principle components or dimensionality to generate in the resulting matrix. Default 20L.

lambda
A numeric scalar. Algorithmic parameter, the penalty parameter which limits the dataset-specific component of the factorization. Default 5.0.

resolution
A numeric scalar. Algorithmic paramter, the clustering resolution, increasing this increases the number of communities detected. Default 1.0

Value

The input SingleCellExperiment object with reducedDim(inSCE, reducedDimName) updated.

References

Joshua Welch, et al., 2018

Examples

if (FALSE) {
data('sceBatches', package = 'singleCellTK')
sceCorr <- runLIGER(sceBatches)
}