runMAST

Run MAST for differential expression analysis on a SingleCellExperiment inherited object.

runMAST(
  inSCE,
  useAssay = "logcounts",
  index1 = NULL,
  index2 = NULL,
  class = NULL,
  classGroup1 = NULL,
  classGroup2 = NULL,
  comparisonName,
  groupName1 = NULL,
  groupName2 = NULL,
  useThresh = FALSE,
  freqExpressed = 0.1,
  onlyPos = FALSE,
  log2fcThreshold = NULL,
  fdrThreshold = 1
)

Arguments

inSCE
SingleCellExperiment inherited object.

useAssay
character. A string specifying which assay to use for the MAST calculations. Default "logcounts".

index1
Any type of indices that can subset a SingleCellExperiment inherited object by cells. Specifies which cells are of interests. Default NULL.

index2
Any type of indices that can subset a SingleCellExperiment inherited object by cells. specifies which cells are being compared with those specified by index1. If NULL, index1 cells will be comapred with all other cells. Default NULL.

class
A vector/factor with ncol(inSCE) elements, or a single string that specifies a column name of colData(inSCE). Default NULL.

classGroup1
a vector specifying which "levels" given in class are of interests. Default NULL.

classGroup2
a vector specifying which "levels" given in class will be compared with those specified by classGroup1. If NULL, classGroup1 cells will be compared with all other cells.

comparisonName
A string naming the DEG experiment. Will be required in downstream plotting functions.

groupName1
A string naming the first comparison group - the condition you are interested in.

groupName2
A string naming the second comparison group - the condition that the first group is being compared with.

useThresh
Whether to use adaptive thresholding to filter genes. Default FALSE.

freqExpressed
A numeric threshold that the genes expressed in less than this fraction of cells will be removed. Default 0.1.

onlyPos
Whether to only output DEG with positive log2_FC value. Default FALSE.

log2fcThreshold
Only out put DEGs with the absolute values of log2FC larger than this value. Default NULL

fdrThreshold
Only out put DEGs with FDR value smaller than this value. Default 1

Value

The input SingleCellExperiment object with metadata(inSCE)$MAST updated with the results: a list named by comparisonName, with $groupNames containing the naming of the two conditions, $useAssay storing the assay name that was used for calculation, $select storing the cell selection indices (logical) for each condition, and $result storing a data.frame of the DEGs summary.

Details

Condition group can be set by one and only one of the two ways: setting index1 and index2, or setting class, classGroup1 and classGroup2.

Examples

data('sceBatches', package = 'singleCellTK')
sce.deg <- runMAST(sceBatches, class = "cell_type", comparisonName = 'aVSb',
                   classGroup1 = 'alpha', classGroup2 = 'beta',
                   groupName1 = 'a', groupName2 = 'b')

#> Error in getClass(element.type): “SimpleAssays” is not a defined class