Wrapper for calculating QC metrics with scater.

A wrapper function for decontX. Identify potential contamination from experimental factors such as ambient RNA.

runPerCellQC(
  inSCE,
  useAssay = "counts",
  geneSetList = NULL,
  geneSetListLocation = "rownames",
  geneSetCollection = NULL,
  ...
)

Arguments

inSCE
Input SingleCellExperiment object.

useAssay
A string specifying which assay in the SCE to use. Default 'counts'.

geneSetList
List of gene sets to be quantified. The genes in the assays will be matched to the genes in the list based on geneSetListLocation.

geneSetListLocation
Character or numeric vector. If set to 'rownames', then the genes in 'geneSetList' will be looked up in rownames(inSCE). If another character is supplied, then genes will be looked up in the column names of rowData(inSCE). A character vector with the same length as geneSetList can be supplied if the IDs for different gene sets are found in different places, including a mixture of 'rownames' and rowData(inSCE). An integer or integer vector can be supplied to denote the column index in rowData(inSCE). Default 'rownames'.

geneSetCollection
Class of GeneSetCollection from package GSEAbase. The location of the gene IDs in inSCE should be in the description slot of each gene set and should follow the same notation as geneSetListLocation. The function getGmt can be used to read in gene sets from a GMT file. If reading a GMT file, the second column for each gene set should be the description denoting the location of the gene IDs in inSCE. These gene sets will be included with those from geneSetList if both parameters are provided.

...
Additional arguments to pass to addPerCellQC.

Value

A SingleCellExperiment object with cell QC metrics added to the colData slot. If geneSetList or geneSetCollection are provided, then the rownames for each gene set will be saved in metadata(inSCE)$scater$addPerCellQC$geneSets.

Examples

data(qcSceExample, package = "singleCellTK")

#> Warning: data set ‘qcSceExample’ not found

mito.ix = grep("^MT-", rowData(sce)$feature_name)

#> Error in `rownames<-`(`*tmp*`, value = c("ENSG00000251562", "ENSG00000205542", "ENSG00000177954", "ENSG00000166710", "ENSG00000156508", "ENSG00000147403", "ENSG00000229117", "ENSG00000167526", "ENSG00000109475", "ENSG00000122026", "ENSG00000142541", "ENSG00000137818", "ENSG00000231500", "ENSG00000198804", "ENSG00000198918", "ENSG00000140988", "ENSG00000213741", "ENSG00000144713", "ENSG00000112306", "ENSG00000075624", "ENSG00000164587", "ENSG00000137154", "ENSG00000177600", "ENSG00000134419", "ENSG00000145425", "ENSG00000087086", "ENSG00000143947", "ENSG00000105640", "ENSG00000198938", "ENSG00000142676", "ENSG00000034510", "ENSG00000105372", "ENSG00000233927", "ENSG00000100316", "ENSG00000161970", "ENSG00000108107", "ENSG00000198712", "ENSG00000149273", "ENSG00000166441", "ENSG00000234745", "ENSG00000019582", "ENSG00000108298", "ENSG00000186468", "ENSG00000147604", "ENSG00000198242", "ENSG00000115268", "ENSG00000142937", "ENSG00000133112", "ENSG00000110700", "ENSG00000182899", "ENSG00000167996", "ENSG00000090382", "ENSG00000163682", "ENSG00000197958", "ENSG00000138326", "ENSG00000071082", "ENSG00000145592", "ENSG00000118181", "ENSG00000174748", "ENSG00000156482", "ENSG00000198034", "ENSG00000130255", "ENSG00000161016", "ENSG00000197756", "ENSG00000198886", "ENSG00000170889", "ENSG00000204287", "ENSG00000105193", "ENSG00000136942", "ENSG00000182774", "ENSG00000163220", "ENSG00000089009", "ENSG00000187514", "ENSG00000198755", "ENSG00000063177", "ENSG00000162244", "ENSG00000008988", "ENSG00000198763", "ENSG00000198840", "ENSG00000171863", "ENSG00000083845", "ENSG00000204525", "ENSG00000143546", "ENSG00000149806", "ENSG00000188846", "ENSG00000172809", "ENSG00000131469", "ENSG00000198899", "ENSG00000170345", "ENSG00000221983", "ENSG00000122406", "ENSG00000204628", "ENSG00000206503", "ENSG00000168028", "ENSG00000148303", "ENSG00000196154", "ENSG00000171858", "ENSG00000124614", "ENSG00000198727", "ENSG00000142534", "ENSG00000171223", "ENSG00000173812", "ENSG00000174444", "ENSG00000116251", "ENSG00000169442", "ENSG00000198888", "ENSG00000089157", "ENSG00000197956", "ENSG00000101439", "ENSG00000132475", "ENSG00000108518", "ENSG00000114391", "ENSG00000196126", "ENSG00000196531", "ENSG00000241343", "ENSG00000070756", "ENSG00000223865", "ENSG00000125691", "ENSG00000184009", "ENSG00000163041", "ENSG00000123349", "ENSG00000124172", "ENSG00000104529", "ENSG00000167658", "ENSG00000135486", "ENSG00000140264", "ENSG00000104904", "ENSG00000231389", "ENSG00000172757", "ENSG00000120129", "ENSG00000051523", "ENSG00000026025", "ENSG00000198786", "ENSG00000142669", "ENSG00000145741", "ENSG00000111640", "ENSG00000011600", "ENSG00000227507", "ENSG00000114942", "ENSG00000169100", "ENSG00000163131", "ENSG00000177606", "ENSG00000133639", "ENSG00000131143", "ENSG00000265681", "ENSG00000156467", "ENSG00000108654", "ENSG00000130066", "ENSG00000092841", "ENSG00000101608", "ENSG00000181163", "ENSG00000204592", "ENSG00000179344", "ENSG00000103187", "ENSG00000111348", "ENSG00000105373", "ENSG00000211592", "ENSG00000196683", "ENSG00000127184", "ENSG00000100097", "ENSG00000102879", "ENSG00000167283", "ENSG00000065978", "ENSG00000150991", "ENSG00000135390", "ENSG00000129824", "ENSG00000245532", "ENSG00000104408", "ENSG00000104894", "ENSG00000163466", "ENSG00000169567", "ENSG00000136156", "ENSG00000196735", "ENSG00000204482", "ENSG00000196262", "ENSG00000232112", "ENSG00000008517", "ENSG00000111229", "ENSG00000271503", "ENSG00000188243", "ENSG00000233276", "ENSG00000162511", "ENSG00000122862", "ENSG00000158869", "ENSG00000128016", "ENSG00000130429", "ENSG00000165502", "ENSG00000067082", "ENSG00000115523", "ENSG00000125534", "ENSG00000197747", "ENSG00000197746", "ENSG00000170315", "ENSG00000279483", "ENSG00000140319", "ENSG00000143933", "ENSG00000100650", "ENSG00000265972", "ENSG00000204472", "ENSG00000133872")): attempt to set 'rownames' on an object with no dimensions

geneSet <- list("Mito"=rownames(sce)[mito.ix])

#> Error in eval(expr, envir, enclos): object 'mito.ix' not found

sce <- runPerCellQC(sce, geneSetList = geneSet)

#> Wed May 27 12:30:32 2020 ... Running 'perCellQCMetrics'

#> Error in runPerCellQC(sce, geneSetList = geneSet): object 'geneSet' not found