Apply scGen batch effect correction method to SingleCellExperiment object

scGen is a generative model to predict single-cell perturbation response across cell types, studies and species. It works by combining variational autoencoders and latent space vector arithmetics for high-dimensional single- cell gene expression data.

runSCGEN(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  cellType = "cell_type",
  nEpochs = 50L,
  assayName = "SCGEN"
)

Arguments

inSCE
SingleCellExperiment inherited object. Required.

useAssay
A single character indicating the name of the assay requiring batch correction. Default "logcounts".

batch
A single character indicating a field in colData that annotates the batches. Default "batch".

cellType
A single character. A string indicating a field in colData(inSCE) that defines different cell types. Default 'cell_type'.

nEpochs
An integer. Algorithmic parameter, the number of epochs to iterate and optimize network weights. Default 50L.

assayName
A single characeter. The name for the corrected assay. Will be saved to assay. Default "SCGEN".

Value

The input SingleCellExperiment object with assay(inSCE, assayName) updated.

References

Lotfollahi, Mohammad et al., 2019

Examples

if (FALSE) {
data('sceBatches', package = 'singleCellTK')
sceCorr <- runSCGEN(sceBatches)
}