Apply scMerge batch effect correction method to SingleCellExperiment object

The scMerge method leverages factor analysis, stably expressed genes (SEGs) and (pseudo-) replicates to remove unwanted variations and merge multiple scRNA-Seq data.

runSCMerge(
  inSCE,
  useAssay = "logcounts",
  batch = "batch",
  assayName = "scMerge",
  seg = NULL,
  kmeansK = NULL,
  cellType = "cell_type",
  nCores = 1
)

Arguments

inSCE
SingleCellExperiment object. An object that stores your dataset and analysis procedures.

useAssay
character, default `"logcounts"`. A string indicating the name of the assay requiring batch correction in "inSCE", should exist in `assayNames(inSCE)`.

batch
character, default `"batch"`. A string indicating the field of `colData(inSCE)` that defines different batches.

assayName
character, default `"scMerge"`. The name for the corrected full-sized expression matrix.

seg
array, default `NULL`. An array of gene names or indices that specifies SEG (Stably Expressed Genes) set as negative control. Pre-defined dataset with human and mouse SEG lists is available to user by running `data('SEG')`.

kmeansK
vector of int, default `NULL`. A vector indicating the kmeans' K-value for each batch, in order to construct pseudo-replicates. The length of `kmeansK` needs to be the same as the number of batches.

cellType
character, default `"cell_type"`. A string indicating the field of `colData(inSCE)` that defines different cell types.

nCores
integer, default `parallel::detectCores()`. The number of cores of processors to allocate for the task. By default it takes all the cores available to the user.

Value

SingleCellExperiment object with `assay(inSCE, assayName)` updated with corrected full-sized expression matrix.

References

Hoa, et al., 2020

Examples

data('sceBatches', package = 'singleCellTK')
sceCorr <- runSCMerge(sceBatches)

#> Error in getClass(element.type): “SimpleAssays” is not a defined class