runSeuratFindClusters Computes the clusters from the input sce object and stores them back in sce object

runSeuratFindClusters Computes the clusters from the input sce object and stores them back in sce object

runSeuratFindClusters(
  inSCE,
  useAssay = "seuratScaledData",
  useReduction = c("pca", "ica"),
  dims = 10,
  algorithm = c("louvain", "multilevel", "SLM"),
  groupSingletons = TRUE,
  resolution = 0.8,
  externalReduction = NULL,
  verbose = TRUE
)

Arguments

inSCE

(sce) object from which clusters should be computed and stored in

useAssay

Assay containing scaled counts to use for clustering.

useReduction

Reduction method to use for computing clusters. One of "pca" or "ica". Default "pca".

dims

numeric value of how many components to use for computing clusters. Default 10.

algorithm

selected algorithm to compute clusters. One of "louvain", "multilevel", or "SLM". Use louvain for "original Louvain algorithm" and multilevel for "Louvain algorithm with multilevel refinement". Default louvain.

groupSingletons

boolean if singletons should be grouped together or not. Default TRUE.

resolution

Set the resolution parameter to find larger (value above 1) or smaller (value below 1) number of communities. Default 0.8.

externalReduction

Pass DimReduc object if PCA/ICA computed through other libraries. Default NULL.

verbose

Logical value indicating if informative messages should be displayed. Default is TRUE.

Value

Updated sce object which now contains the computed clusters

Examples

data(scExample, package = "singleCellTK")
if (FALSE) {
sce <- runSeuratNormalizeData(sce, useAssay = "counts")
sce <- runSeuratFindHVG(sce, useAssay = "counts")
sce <- runSeuratScaleData(sce, useAssay = "counts")
sce <- runSeuratPCA(sce, useAssay = "counts")
sce <- runSeuratFindClusters(sce, useAssay = "counts")
}