saveBiomarkerRes Save biomarker gene information with a custom name when provided with diffex results.

saveBiomarkerRes Save biomarker gene information with a custom name when provided with diffex results.

saveBiomarkerRes(
  inSCE,
  diffex,
  biomarkerName,
  method,
  ntop = 25,
  logFC = NULL,
  pVal = NULL
)

Arguments

inSCE

Input SCtkExperiment object. Required

diffex

results table saved from the differential expression analysis. Required.

biomarkerName

name of the biomarker result to be saved under in rowData(). Required.

method

name of the diffex method used to generate the results. Options are DESeq2, Limma and ANOVA. Required

ntop

number of top N genes. Default is 25. Required

logFC

logfold-change cutoff applied to save biomarker results. Optional

pVal

adjusted p-value cutoff. Optional

Value

a new SCE object with the diffex result saved in the rowData using the "biomarkerName"

Examples

data("mouseBrainSubsetSCE")
res <- scDiffEx(mouseBrainSubsetSCE,
                useAssay = "logcounts",
                condition = "level1class",
                ntop = length(rownames(mouseBrainSubsetSCE)),
                usesig = FALSE,
                diffexmethod = "limma")
#> Warning: Zero sample variances detected, have been offset away from zero
#> Removing intercept from test coefficients
sceObj <- saveDiffExResults(mouseBrainSubsetSCE, res, "Limma_Level1class", "limma")
bioMarkRes <- saveBiomarkerRes(sceObj, res, "bioMarker", "limma", logFC = 4, pVal = 0.05)