Perform differential expression analysis with DESeq2

Returns a data frame of gene names and adjusted p-values

scDiffEx_deseq2(inSCESet, use_assay = "counts", condition,
  analysis_type = "biomarker", levelofinterest = NULL,
  controlLevel = NULL, covariates = NULL)

Arguments

inSCESet
Input SCtkExperiment object. Required

use_assay
Indicate which assay to use. Default is "counts"

condition
The name of the condition to use for differential expression. Must be a name of a column from colData that contains at least two labels. Required

analysis_type
Choose "biomarker" to compare the levelofinterest to all other samples. Choose "contrast" to compare the levelofinterest to a controlLevel (see below).

levelofinterest
If the condition has more than two labels, levelofinterest should contain one factor of interest from condition.

controlLevel
If the condition has more than two labels, controlLevel should contain one factor from condition to use as the control.

covariates
Additional covariates to add to the model.

Value

A data frame of gene names and adjusted p-values

Examples

data("GSE60361_subset_sce")
#subset to 100 genes
subset <- GSE60361_subset_sce[rownames(GSE60361_subset_sce)[order(rowSums(assay(GSE60361_subset_sce, "counts")), decreasing = TRUE)][1:100], ]
res <- scDiffEx_deseq2(subset, condition = "level1class")

#> estimating size factors

#> estimating dispersions

#> gene-wise dispersion estimates

#> mean-dispersion relationship

#> -- note: fitType='parametric', but the dispersion trend was not well captured by the
#>    function: y = a/x + b, and a local regression fit was automatically substituted.
#>    specify fitType='local' or 'mean' to avoid this message next time.

#> final dispersion estimates

#> fitting model and testing