Installs Python packages into a Conda environment

Install all Python packages used in the singleCellTK package using conda_install from package reticulate. This will create a new Conda environment with the name envname if not already present. Note that Anaconda or Miniconda already need to be installed on the local system.

sctkPythonInstallConda(
  envname = "sctk-reticulate",
  conda = "auto",
  packages = c("scipy", "numpy", "astroid", "six"),
  pipPackages = c("scrublet", "scanpy", "bbknn", "scanorama", "anndata"),
  selectConda = TRUE,
  forge = FALSE,
  pipIgnoreInstalled = TRUE,
  pythonVersion = NULL,
  ...
)

Arguments

envname
Character. Name of the conda environment to create.

conda
Character. Path to conda executable. Usue "auto" to find conda using the PATH and other conventional install locations. Default 'auto'.

packages
Character Vector. List of packages to install from Conda.

pipPackages
Character Vector. List of packages to install into the Conda environment using 'pip'.

selectConda
Boolean. Run selectSCTKConda after installing all packages to select the Conda environment. Default TRUE.

forge
Boolean. Include the Conda Forge repository.

pipIgnoreInstalled
Boolean. Ignore installed versions when using pip. This is TRUE by default so that specific package versions can be installed even if they are downgrades. The FALSE option is useful for situations where you don't want a pip install to attempt an overwrite of a conda binary package (e.g. SciPy on Windows which is very difficult to install via pip due to compilation requirements).

pythonVersion
Passed to python_version variable in conda_install. Default NULL.

...
Other parameters to pass to conda_install.

Value

None. Installation of Conda environment.

See also

Examples

if (FALSE) {
sctkPythonInstallConda(envname = "sctk-reticulate")
}