Installs Python packages into a virtual environment

Install all Python packages used in the singleCellTK package using virtualenv_install from package reticulate. This will create a new virtual environment with the name envname if not already present.

sctkPythonInstallVirtualEnv(
  envname = "sctk-reticulate",
  packages = c("scipy", "numpy", "astroid", "six", "scrublet", "scanpy", "scanorama",
    "bbknn", "scgen"),
  selectEnvironment = TRUE,
  python = NULL
)

Arguments

envname
Character. Name of the virtual environment to create.

packages
Character Vector. List of packages to install.

selectEnvironment
Boolean. Run selectSCTKVirtualEnvironment after installing all packages to select the virtual environment. Default TRUE.

python
The path to a Python interpreter, to be used with the created virtual environment. When NULL, the Python interpreter associated with the current session will be used. Default NULL.

See also

Examples

if (FALSE) {
sctkPythonInstallVirtualEnv(envname = "sctk-reticulate")
}