seuratPCA Computes PCA on the input sce object and stores the calculated principal components within the sce object

seuratPCA Computes PCA on the input sce object and stores the calculated principal components within the sce object

seuratPCA(
  inSCE,
  useAssay = "seuratScaledData",
  reducedDimName = "seuratPCA",
  nPCs = 20,
  verbose = TRUE
)

Arguments

inSCE
(sce) object on which to compute PCA

useAssay
Assay containing scaled counts to use in PCA.

reducedDimName
Name of new reducedDims object containing Seurat PCA. Default seuratPCA.

nPCs
numeric value of how many components to compute. Default 20.

verbose
Logical value indicating if informative messages should be displayed. Default is TRUE.

Value

Updated SingleCellExperiment object which now contains the computed principal components

Examples

data(scExample, package = "singleCellTK")
if (FALSE) {
sce <- seuratNormalizeData(sce, useAssay = "counts")
sce <- seuratFindHVG(sce, useAssay = "counts")
sce <- seuratScaleData(sce, useAssay = "counts")
sce <- seuratPCA(sce, useAssay = "counts")
}