Computes an HTML report from the Seurat workflow and returns the output SCE object with the computations stored in it.

Computes an HTML report from the Seurat workflow and returns the output SCE object with the computations stored in it.

seuratReport(
  inSCE,
  outputFile = NULL,
  outputDir = NULL,
  subtitle = "BUMC Single Cell Sequencing Core",
  authors = "Tianmu (Timo) Hu, Irzam Sarfraz",
  sce = NULL,
  biological.group = NULL,
  phenotype.groups = NULL,
  selected.markers = NULL,
  clustering.resolution = 0.8,
  variable.features = 2000,
  pc.count = 10,
  showSession = TRUE,
  pdf = TRUE
)

Arguments

inSCE
Input SingleCellExperiment object.

outputFile
Specify the name of the generated output HTML file. If NULL then the output file name will be based on the name of the Rmarkdown template. Default NULL.

outputDir
Specify the name of the output directory to save the rendered HTML file. If NULL the file is stored to the current working directory.

subtitle
A character value specifying the subtitle to use in the Seurat report.

authors
A character value specifying the names of the authors to use in the Seurat report.

sce
A character value specifying the path of the input SingleCellExperiment object.

biological.group
A character value that specifies the name of the colData column to use as the main biological group in the seurat report for differential expression and grouping.

phenotype.groups
A character vector that specifies the names of the colData columns to use for differential expression in addition to the biological.group parameter.

selected.markers
A character vector specifying the user decided gene symbols of pre-selected markers that be used to generate gene plots in addition to the gene markers computed from differential expression.

clustering.resolution
A numeric value indicating the resolution to use with clustering. Default is 0.8.

variable.features
A numeric value indicating the number of top variable genes to identify in the seurat report. Default is 2000.

pc.count
A numeric value indicating the number of principal components to use in the analysis workflow. Default is 10.

showSession
A logical value indicating if session information should be displayed or not. Default is TRUE.

pdf
A logical value indicating if a pdf should also be generated for each figure in the report. Default is TRUE.

Value

A SingleCellExperiment object that has the seurat computations stored and can be used to interactively visualize the plots by importing in the singleCellTK user interface.