seuratRunUMAP Computes UMAP from the given sce object and stores the UMAP computations back into the sce object

seuratRunUMAP Computes UMAP from the given sce object and stores the UMAP computations back into the sce object

seuratRunUMAP(
  inSCE,
  useReduction = c("pca", "ica"),
  reducedDimName = "seuratUMAP",
  dims = 10,
  minDist = 0.3,
  nNeighbors = 30L,
  spread = 1
)

Arguments

inSCE
(sce) object on which to compute the UMAP

useReduction
Reduction to use for computing UMAP. One of "pca" or "ica". Default is "pca".

reducedDimName
Name of new reducedDims object containing Seurat UMAP Default seuratUMAP.

dims
Numerical value of how many reduction components to use for UMAP computation. Default 10.

minDist
Sets the "min.dist" parameter to the underlying UMAP call. See RunUMAP for more information. Default 0.3.

nNeighbors
Sets the "n.neighbors" parameter to the underlying UMAP call. See RunUMAP for more information. Default 30L.

spread
Sets the "spread" parameter to the underlying UMAP call. See RunUMAP for more information. Default 1.

Value

Updated sce object with UMAP computations stored