MAST

Run and visualize MAST analysis on a SCtkExperiment object.

MAST(inSCE, condition = NULL, interest.level = NULL,
  freqExpressed = 0.1, fcThreshold = log2(1.5), p.value = 0.05,
  useThresh = FALSE, useAssay = "logcounts")

thresholdGenes(inSCE, useAssay = "logcounts")

MASTviolin(inSCE, useAssay = "logcounts", fcHurdleSig, samplesize = 49,
  threshP = FALSE, condition)

MASTregression(inSCE, useAssay = "logcounts", fcHurdleSig,
  samplesize = 49, threshP = FALSE, condition)

Arguments

inSCE
Input SCtkExperiment object. Required

condition
select variable (from the colData) that is used for the model.

interest.level
If the condition of interest has more than two factors, indicate which level should be used to compare to all other samples.

freqExpressed
Filter genes that are expressed in at least this fraction of cells. The default is expression in 0.1 of samples.

fcThreshold
Minimum fold change for differentially expressed gene.

p.value
p values for selecting the hurdle result, default is 0.05

useThresh
Use adaptive thresholding to filter genes. The default is FALSE.

useAssay
The assay to use for the MAST calculations. The default is "logcounts"

fcHurdleSig
The filtered result from hurdle model

samplesize
The number of most significant genes

threshP
Plot threshold values from adaptive thresholding. Default is FALSE

Value

MAST(): A data.frame of differentially expressed genes with p-values.

thresholdGenes(): list of thresholded counts (on natural scale), thresholds, bins, densities estimated on each bin, and the original data from MAST::thresholdSCRNACountMatrix

MASTviolin(): A ggplot object of MAST violin plots.

MASTregression(): A ggplot object of MAST linear regression plots.

Functions

MAST: Run MAST analysis.

thresholdGenes: Identify adaptive thresholds

MASTviolin: Visualize MAST results using violin plots

MASTregression: Visualize MAST results using linear model plots

Examples

data("mouseBrainSubsetSCE")
res <- thresholdGenes(mouseBrainSubsetSCE)

#> `fData` has no primerid.  I'll make something up.

#> `cData` has no wellKey.  I'll make something up.

#> Assuming data assay in position 1, with name et is log-transformed.