Align Single Cell RNA-Seq Data and Create a SCESet Object

Align Single Cell RNA-Seq Data and Create a SCESet Object

alignSingleCellData(inputfile1, inputfile2 = NULL, index_path, gtf_annotation,
  output_dir = NULL, sample_annotations = NULL,
  feature_annotations = NULL, threads = 1, save_bam = FALSE,
  save_count_files = FALSE, isPairedEnd = FALSE)

Arguments

inputfile1
An input file or list of files. Files can be fastq, fastq.gz, or bam, but must all be of the same type. Sample names will be the full file name, without _1.fastq.gz, .fastq.gz, _1.fastq, .fastq or .bam endings.

inputfile2
If fastq files are provided in input list, a list of corresponding paired fastq files, if applicable.

index_path
Path to the Rsubread genome index.

gtf_annotation
Path to the GTF gene annotation to use. This must correspond to the genome specified in index_path.

output_dir
If save_bam or save_count_files is TRUE, specify a directory in which to save the output files.

sample_annotations
A data.frame of sample annotations, with samples as rows and annotations in columns. The sample names must be identical to and in the same order as the list of files in inputfile1. Alignment statistics will be added to the annotation data frame.

feature_annotations
An optional data.frame of probe annotations, with probes as rows and probe annotations in columns.

threads
Number of threads to use during alignment. The default is 1.

save_bam
If TRUE, bam alignment files will be saved in the output_dir. The default is FALSE.

save_count_files
If TRUE, per sample gene count files will be saved in the output_dir. The default is FALSE.

isPairedEnd
If input files are .bam, indicate whether the input bam files are paired end.

Value

Object to import into the shiny app.

Examples

# NOT RUN {
singlecellobject <- alignSingleCellData(
  inputfile1 = c("/path/to/sample1_1.fastq.gz",
                 "/path/to/sample2_1.fastq.gz"),
  inputfile2 = c("/path/to/sample1_2.fastq.gz",
                 "/path/to/sample2_2.fastq.gz"),
  index_path = "/path/to/genome/index",
  gtf_annotation = "/path/to/gene/annotations.gtf",
  sample_annotations = sample.annotation.df,
  threads=4)
# }