Align Single Cell RNA-Seq Data and Create a SCtkExperiment Object

Align Single Cell RNA-Seq Data and Create a SCtkExperiment Object

alignSingleCellData(inputfile1, inputfile2 = NULL, indexPath,
  gtfAnnotation, outputDir = NULL, sampleAnnotations = NULL,
  featureAnnotations = NULL, threads = 1, saveBam = FALSE,
  saveCountFiles = FALSE, isPairedEnd = FALSE)

Arguments

inputfile1
An input file or list of files. Files can be fastq, fastq.gz, or bam, but must all be of the same type. Sample names will be the full file name, without _1.fastq.gz, .fastq.gz, _1.fastq, .fastq or .bam endings.

inputfile2
If fastq files are provided in input list, a list of corresponding paired fastq files, if applicable.

indexPath
Path to the Rsubread genome index.

gtfAnnotation
Path to the GTF gene annotation to use. This must correspond to the genome specified in indexPath.

outputDir
If saveBam or saveCountFiles is TRUE, specify a directory in which to save the output files.

sampleAnnotations
A data.frame of sample annotations, with samples as rows and annotations in columns. The sample names must be identical to and in the same order as the list of files in inputfile1. Alignment statistics will be added to the annotation data frame.

featureAnnotations
An optional data.frame of probe annotations, with probes as rows and probe annotations in columns.

threads
Number of threads to use during alignment. The default is 1.

saveBam
If TRUE, bam alignment files will be saved in the outputDir. The default is FALSE.

saveCountFiles
If TRUE, per sample gene count files will be saved in the outputDir. The default is FALSE.

isPairedEnd
If input files are .bam, indicate whether the input bam files are paired end.

Value

Object to import into the shiny app.

Examples

# NOT RUN {
singlecellobject <- alignSingleCellData(
  inputfile1 = c("/path/to/sample1_1.fastq.gz",
                 "/path/to/sample2_1.fastq.gz"),
  inputfile2 = c("/path/to/sample1_2.fastq.gz",
                 "/path/to/sample2_2.fastq.gz"),
  indexPath = "/path/to/genome/index",
  gtfAnnotation = "/path/to/gene/annotations.gtf",
  sampleAnnotations = sample.annotation.df,
  threads=4)
# }