Convert Gene IDs

Convert the gene IDs in a SingleCellExperiment object using Bioconductor org.*.eg.db data packages. Because annotation databases do not have a 1:1 relationship, this tool removes rows with no corresponding annotation in your desired annotation, and remove any duplicate annotations after conversion.

convertGeneIDs(inSCE, inSymbol, outSymbol, database = "org.Hs.eg.db")

Arguments

inSCE
Input SingleCellExperiment object.

inSymbol
The input symbol type

outSymbol
The output symbol type

database
The org.*.eg.db database to use. The default is org.Hs.eg.db

Value

A SingleCellExperiment object with converted gene IDs.

Examples

if(requireNamespace("org.Mm.eg.db", quietly = TRUE)) {
  #convert mouse gene symbols to ensembl IDs
  library("org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSCE), 50)
  mouseBrainSubsetSymbol <- convertGeneIDs(inSCE = mouseBrainSubsetSCE,
                                           inSymbol = "SYMBOL",
                                           outSymbol = "ENSEMBL",
                                           database = "org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSymbol), 50)
}

#> 

#> Loading required package: AnnotationDbi

#> 'select()' returned 1:many mapping between keys and columns

#> 109 ENSEMBL are NA after conversion. Removing

#> 34 SYMBOL are duplicated after conversion. Removing additional copies

#> 2314 SYMBOL originally, 2205 ENSEMBLs

#>  [1] "ENSMUSG00000028719" "ENSMUSG00000050854" "ENSMUSG00000026867"
#>  [4] "ENSMUSG00000026853" "ENSMUSG00000068039" "ENSMUSG00000026234"
#>  [7] "ENSMUSG00000030824" "ENSMUSG00000094870" "ENSMUSG00000033379"
#> [10] "ENSMUSG00000019362" "ENSMUSG00000094483" "ENSMUSG00000046985"
#> [13] "ENSMUSG00000039361" "ENSMUSG00000026204" "ENSMUSG00000024597"
#> [16] "ENSMUSG00000022337" "ENSMUSG00000021024" "ENSMUSG00000036402"
#> [19] "ENSMUSG00000050121" "ENSMUSG00000006800" "ENSMUSG00000033184"
#> [22] "ENSMUSG00000005312" "ENSMUSG00000037826" "ENSMUSG00000003974"
#> [25] "ENSMUSG00000073676" "ENSMUSG00000028568" "ENSMUSG00000040249"
#> [28] "ENSMUSG00000033004" "ENSMUSG00000015850" "ENSMUSG00000028484"
#> [31] "ENSMUSG00000029461" "ENSMUSG00000000131" "ENSMUSG00000063229"
#> [34] "ENSMUSG00000050708" "ENSMUSG00000032026" "ENSMUSG00000066551"
#> [37] "ENSMUSG00000006005" "ENSMUSG00000020831" "ENSMUSG00000004151"
#> [40] "ENSMUSG00000054162" "ENSMUSG00000027642" "ENSMUSG00000009013"
#> [43] "ENSMUSG00000047675" "ENSMUSG00000015961" "ENSMUSG00000022283"
#> [46] "ENSMUSG00000000346" "ENSMUSG00000037062" "ENSMUSG00000060703"
#> [49] "ENSMUSG00000029405" "ENSMUSG00000027523"