Convert Gene IDs

Convert the gene IDs in a SingleCellExperiment object using Bioconductor org.*.eg.db data packages. Because annotation databases do not have a 1:1 relationship, this tool removes rows with no corresponding annotation in your desired annotation, and remove any duplicate annotations after conversion.

convertGeneIDs(inSCE, inSymbol, outSymbol, database = "org.Hs.eg.db")

Arguments

inSCE
Input SCtkExperiment object. Required

inSymbol
The input symbol type

outSymbol
The output symbol type

database
The org.*.eg.db database to use. The default is org.Hs.eg.db

Value

A SCtkExperiment with converted gene IDs.

Examples

if(requireNamespace("org.Mm.eg.db", quietly = TRUE)) {
  #convert mouse gene symbols to ensembl IDs
  library("org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSCE), 50)
  mouseBrainSubsetSymbol <- convertGeneIDs(inSCE = mouseBrainSubsetSCE,
                                           inSymbol = "SYMBOL",
                                           outSymbol = "ENSEMBL",
                                           database = "org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSymbol), 50)
}

#> 

#> Loading required package: AnnotationDbi

#> 'select()' returned 1:many mapping between keys and columns

#> 1924 ENSEMBL are NA after conversion. Removing

#> 39 ENSEMBL are duplicated after conversion. Removing additional copies

#> 332 SYMBOL are duplicated after conversion. Removing additional copies

#> 19972 SYMBOL originally, 18013 ENSEMBLs

#>  [1] "ENSMUSG00000002058" "ENSMUSG00000034839" "ENSMUSG00000029344"
#>  [4] "ENSMUSG00000026873" "ENSMUSG00000035000" "ENSMUSG00000024011"
#>  [7] "ENSMUSG00000114469" "ENSMUSG00000053129" "ENSMUSG00000031023"
#> [10] "ENSMUSG00000037706" "ENSMUSG00000021745" "ENSMUSG00000066551"
#> [13] "ENSMUSG00000017418" "ENSMUSG00000049422" "ENSMUSG00000001473"
#> [16] "ENSMUSG00000051910" "ENSMUSG00000029110" "ENSMUSG00000061013"
#> [19] "ENSMUSG00000074345" "ENSMUSG00000019842" "ENSMUSG00000068252"
#> [22] "ENSMUSG00000025784" "ENSMUSG00000031875" "ENSMUSG00000102046"
#> [25] "ENSMUSG00000055780" "ENSMUSG00000021552" "ENSMUSG00000028291"
#> [28] "ENSMUSG00000024712" "ENSMUSG00000024966" "ENSMUSG00000042312"
#> [31] "ENSMUSG00000026283" "ENSMUSG00000048039" "ENSMUSG00000025464"
#> [34] "ENSMUSG00000054717" "ENSMUSG00000024049" "ENSMUSG00000073000"
#> [37] "ENSMUSG00000061136" "ENSMUSG00000041923" "ENSMUSG00000073103"
#> [40] "ENSMUSG00000030505" "ENSMUSG00000028687" "ENSMUSG00000029848"
#> [43] "ENSMUSG00000079262" "ENSMUSG00000053553" "ENSMUSG00000064307"
#> [46] "ENSMUSG00000031661" "ENSMUSG00000049456" "ENSMUSG00000009941"
#> [49] "ENSMUSG00000054537" "ENSMUSG00000029445"