Convert Gene IDs

Convert the gene IDs in a SingleCellExperiment object using Bioconductor org.*.eg.db data packages. Because annotation databases do not have a 1:1 relationship, this tool removes rows with no corresponding annotation in your desired annotation, and remove any duplicate annotations after conversion.

convertGeneIDs(inSCESet, inSymbol, outSymbol, database = "org.Hs.eg.db")

Arguments

inSCESet
Input SCtkExperiment object. Required

inSymbol
The input symbol type

outSymbol
The output symbol type

database
The org.*.eg.db database to use. The default is org.Hs.eg.db

Value

A SCtkExperiment with converted gene IDs.

Examples

if(requireNamespace("org.Mm.eg.db", quietly = TRUE)) {
  #convert mouse gene symbols to ensembl IDs
  library("org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSCE), 50)
  mouseBrainSubsetSymbol <- convertGeneIDs(inSCESet = mouseBrainSubsetSCE,
                                              inSymbol = "SYMBOL",
                                              outSymbol = "ENSEMBL",
                                              database = "org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSymbol), 50)
}

#> 

#> Loading required package: AnnotationDbi

#> 'select()' returned 1:many mapping between keys and columns

#> 1773 ENSEMBL are NA after conversion. Removing

#> 51 ENSEMBL are duplicated after conversion. Removing additional copies

#> 153 SYMBOL are duplicated after conversion. Removing additional copies

#> 19972 SYMBOL originally, 18153 ENSEMBLs

#>  [1] "ENSMUSG00000026818" "ENSMUSG00000025184" "ENSMUSG00000043843"
#>  [4] "ENSMUSG00000024583" "ENSMUSG00000029003" "ENSMUSG00000027306"
#>  [7] "ENSMUSG00000028332" "ENSMUSG00000020061" "ENSMUSG00000052595"
#> [10] "ENSMUSG00000018500" "ENSMUSG00000029075" "ENSMUSG00000028282"
#> [13] "ENSMUSG00000022711" "ENSMUSG00000053553" "ENSMUSG00000027206"
#> [16] "ENSMUSG00000052056" "ENSMUSG00000029212" "ENSMUSG00000047562"
#> [19] "ENSMUSG00000071860" "ENSMUSG00000062393" "ENSMUSG00000028212"
#> [22] "ENSMUSG00000030096" "ENSMUSG00000113450" "ENSMUSG00000019838"
#> [25] "ENSMUSG00000065452" "ENSMUSG00000108356" "ENSMUSG00000070284"
#> [28] "ENSMUSG00000093411" "ENSMUSG00000054752" "ENSMUSG00000055403"
#> [31] "ENSMUSG00000031953" "ENSMUSG00000001674" "ENSMUSG00000074881"
#> [34] "ENSMUSG00000048118" "ENSMUSG00000043924" "ENSMUSG00000058325"
#> [37] "ENSMUSG00000007892" "ENSMUSG00000029679" "ENSMUSG00000036816"
#> [40] "ENSMUSG00000001415" "ENSMUSG00000031216" "ENSMUSG00000020940"
#> [43] "ENSMUSG00000030917" "ENSMUSG00000092203" "ENSMUSG00000025531"
#> [46] "ENSMUSG00000067049" "ENSMUSG00000087107" "ENSMUSG00000050866"
#> [49] "ENSMUSG00000070583" "ENSMUSG00000054702"