Convert Gene IDs

Convert the gene IDs in a SingleCellExperiment object using Bioconductor org.*.eg.db data packages. Because annotation databases do not have a 1:1 relationship, this tool removes rows with no corresponding annotation in your desired annotation, and remove any duplicate annotations after conversion.

convertGeneIDs(inSCE, inSymbol, outSymbol, database = "org.Hs.eg.db")

Arguments

inSCE
Input SCtkExperiment object. Required

inSymbol
The input symbol type

outSymbol
The output symbol type

database
The org.*.eg.db database to use. The default is org.Hs.eg.db

Value

A SCtkExperiment with converted gene IDs.

Examples

if(requireNamespace("org.Mm.eg.db", quietly = TRUE)) {
  #convert mouse gene symbols to ensembl IDs
  library("org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSCE), 50)
  mouseBrainSubsetSymbol <- convertGeneIDs(inSCE = mouseBrainSubsetSCE,
                                           inSymbol = "SYMBOL",
                                           outSymbol = "ENSEMBL",
                                           database = "org.Mm.eg.db")
  sample(rownames(mouseBrainSubsetSymbol), 50)
}

#> 

#> Loading required package: AnnotationDbi

#> 'select()' returned 1:many mapping between keys and columns

#> 1783 ENSEMBL are NA after conversion. Removing

#> 39 ENSEMBL are duplicated after conversion. Removing additional copies

#> 292 SYMBOL are duplicated after conversion. Removing additional copies

#> 19972 SYMBOL originally, 18152 ENSEMBLs

#>  [1] "ENSMUSG00000014873" "ENSMUSG00000025008" "ENSMUSG00000003378"
#>  [4] "ENSMUSG00000054008" "ENSMUSG00000059423" "ENSMUSG00000019942"
#>  [7] "ENSMUSG00000078713" "ENSMUSG00000091002" "ENSMUSG00000074913"
#> [10] "ENSMUSG00000042189" "ENSMUSG00000065990" "ENSMUSG00000040455"
#> [13] "ENSMUSG00000039427" "ENSMUSG00000058291" "ENSMUSG00000027233"
#> [16] "ENSMUSG00000034792" "ENSMUSG00000027574" "ENSMUSG00000005800"
#> [19] "ENSMUSG00000103472" "ENSMUSG00000087252" "ENSMUSG00000028240"
#> [22] "ENSMUSG00000051065" "ENSMUSG00000032999" "ENSMUSG00000075225"
#> [25] "ENSMUSG00000028264" "ENSMUSG00000052656" "ENSMUSG00000039952"
#> [28] "ENSMUSG00000029787" "ENSMUSG00000039653" "ENSMUSG00000030082"
#> [31] "ENSMUSG00000031949" "ENSMUSG00000042581" "ENSMUSG00000027259"
#> [34] "ENSMUSG00000044712" "ENSMUSG00000037366" "ENSMUSG00000030986"
#> [37] "ENSMUSG00000001134" "ENSMUSG00000029517" "ENSMUSG00000043259"
#> [40] "ENSMUSG00000004896" "ENSMUSG00000048826" "ENSMUSG00000053580"
#> [43] "ENSMUSG00000039911" "ENSMUSG00000043716" "ENSMUSG00000034732"
#> [46] "ENSMUSG00000014426" "ENSMUSG00000020532" "ENSMUSG00000100717"
#> [49] "ENSMUSG00000052912" "ENSMUSG00000042589"