Convert gene IDs

Convert the gene IDs in a SingleCellExperiment object using Bioconductor org.*.eg.db data packages. Because annotation databases do not have a 1:1 relationship, this tool removes rows with no corresponding annotation in your desired annotation, and remove any duplicate annotations after conversion.

convert_gene_ids(inSCESet, in_symbol, out_symbol, database = "org.Hs.eg.db")

Arguments

inSCESet
Input SCtkExperiment object. Required

in_symbol
The input symbol type

out_symbol
The output symbol type

database
The org.*.eg.db database to use. The default is org.Hs.eg.db

Value

A SCtkExperiment with converted gene IDs.

Examples

if(requireNamespace("org.Mm.eg.db", quietly = TRUE)) {
  #convert mouse gene symbols to ensembl IDs
  library("org.Mm.eg.db")
  sample(rownames(GSE60361_subset_sce), 50)
  GSE60361_subset_symbol <- convert_gene_ids(inSCESet = GSE60361_subset_sce,
                                             in_symbol = "SYMBOL",
                                             out_symbol = "ENSEMBL",
                                             database = "org.Mm.eg.db")
  sample(rownames(GSE60361_subset_symbol), 50)
}

#> 

#> 'select()' returned 1:many mapping between keys and columns

#> 1773 ENSEMBL are NA after conversion. Removing

#> 51 ENSEMBL are duplicated after conversion. Removing additional copies

#> 153 SYMBOL are duplicated after conversion. Removing additional copies

#> 19972 SYMBOL originally, 18153 ENSEMBLs

#>  [1] "ENSMUSG00000036282" "ENSMUSG00000033416" "ENSMUSG00000024781"
#>  [4] "ENSMUSG00000014773" "ENSMUSG00000061360" "ENSMUSG00000029601"
#>  [7] "ENSMUSG00000052821" "ENSMUSG00000039611" "ENSMUSG00000067578"
#> [10] "ENSMUSG00000042331" "ENSMUSG00000030351" "ENSMUSG00000033159"
#> [13] "ENSMUSG00000039616" "ENSMUSG00000095138" "ENSMUSG00000067571"
#> [16] "ENSMUSG00000043953" "ENSMUSG00000032059" "ENSMUSG00000004864"
#> [19] "ENSMUSG00000024045" "ENSMUSG00000055013" "ENSMUSG00000021681"
#> [22] "ENSMUSG00000018068" "ENSMUSG00000004415" "ENSMUSG00000022217"
#> [25] "ENSMUSG00000050201" "ENSMUSG00000040724" "ENSMUSG00000020878"
#> [28] "ENSMUSG00000039154" "ENSMUSG00000038670" "ENSMUSG00000020929"
#> [31] "ENSMUSG00000086899" "ENSMUSG00000024806" "ENSMUSG00000067851"
#> [34] "ENSMUSG00000033684" "ENSMUSG00000022243" "ENSMUSG00000031751"
#> [37] "ENSMUSG00000112663" "ENSMUSG00000063605" "ENSMUSG00000001865"
#> [40] "ENSMUSG00000030785" "ENSMUSG00000031434" "ENSMUSG00000056091"
#> [43] "ENSMUSG00000022791" "ENSMUSG00000045252" "ENSMUSG00000045672"
#> [46] "ENSMUSG00000022505" "ENSMUSG00000059208" "ENSMUSG00000063802"
#> [49] "ENSMUSG00000025665" "ENSMUSG00000021983"