Find the marker gene set for each cluster With an input SingleCellExperiment object and specifying the clustering labels, this function iteratively call the differential expression analysis on each cluster against all the others.

Find the marker gene set for each cluster With an input SingleCellExperiment object and specifying the clustering labels, this function iteratively call the differential expression analysis on each cluster against all the others.

findMarkerDiffExp(
  inSCE,
  useAssay = "logcounts",
  method = "MAST",
  cluster = "cluster",
  covariates = NULL,
  log2fcThreshold = 0.25,
  fdrThreshold = 0.05
)

Arguments

inSCE
SingleCellExperiment inherited object.

useAssay
character. A string specifying which assay to use for the MAST calculations. Default "logcounts".

method
A single character for specific differential expression analysis method. Choose from 'MAST', 'DESeq2', 'Limma', 'ANOVA'. Default "MAST".

cluster
One single character to specify a column in colData(inSCE) for the clustering label. Alternatively, a vector or a factor is also acceptable. Default "cluster".

covariates
A character vector of additional covariates to use when building the model. All covariates must exist in names(colData(inSCE)). Not applicable when method is "MAST" method. Default NULL.

log2fcThreshold
Only out put DEGs with the absolute values of log2FC larger than this value. Default NULL

fdrThreshold
Only out put DEGs with FDR value smaller than this value. Default 1

Value

The input SingleCellExperiment object with metadata(inSCE)$findMarker updated with a data.table of the up- regulated DEGs for each cluster.