Given a list of genes and a SCtkExperiment object, return the binary or continuous expression of the genes.

Given a list of genes and a SCtkExperiment object, return the binary or continuous expression of the genes.

getBiomarker(inSCE, gene, binary = "Binary", useAssay = "counts")

Arguments

inSCE
Input SCtkExperiment object. Required

gene
gene list

binary
"Binary" for binary expression or "Continuous" for a gradient. Default: "Binary"

useAssay
Indicate which assay to use. The default is "counts".

Value

getBiomarker(): A data.frame of expression values

Examples

getBiomarker(mouseBrainSubsetSCE, gene="C1qa")

#>                          sample  C1qa
#> X1772063061_G11 X1772063061_G11 FALSE
#> X1772067064_C10 X1772067064_C10 FALSE
#> X1772062113_D09 X1772062113_D09 FALSE
#> X1772062111_A12 X1772062111_A12 FALSE
#> X1772071040_C12 X1772071040_C12 FALSE
#> X1772062111_D09 X1772062111_D09 FALSE
#> X1772071036_G02 X1772071036_G02 FALSE
#> X1772071040_D04 X1772071040_D04 FALSE
#> X1772058177_E12 X1772058177_E12 FALSE
#> X1772066070_G10 X1772066070_G10 FALSE
#> X1772063074_G10 X1772063074_G10 FALSE
#> X1772062115_E07 X1772062115_E07 FALSE
#> X1772062109_F01 X1772062109_F01 FALSE
#> X1772060226_G05 X1772060226_G05 FALSE
#> X1772066108_D06 X1772066108_D06 FALSE
#> X1772060224_F07 X1772060224_F07  TRUE
#> X1772062111_F12 X1772062111_F12  TRUE
#> X1772063078_G02 X1772063078_G02  TRUE
#> X1772060240_H11 X1772060240_H11  TRUE
#> X1772067094_C07 X1772067094_C07  TRUE
#> X1772058177_A09 X1772058177_A09  TRUE
#> X1772058171_C05 X1772058171_C05  TRUE
#> X1772062128_B03 X1772062128_B03  TRUE
#> X1772062118_F04 X1772062118_F04  TRUE
#> X1772062109_E05 X1772062109_E05  TRUE
#> X1772062128_D02 X1772062128_D02  TRUE
#> X1772062109_E12 X1772062109_E12  TRUE
#> X1772062113_E05 X1772062113_E05  TRUE
#> X1772062128_F05 X1772062128_F05  TRUE
#> X1772062118_D09 X1772062118_D09  TRUE