Shows MSigDB categories

Returns a data.frame that shows MSigDB categories and subcategories as well as descriptions for each. The entries in the ID column in this table can be used as input for importGeneSetsFromMSigDB.

getMSigDBTable()

Value

data.frame, containing MSigDB categories

See also

Author

Joshua D. Campbell

Examples

getMSigDBTable()
#>                ID Category Subcategory
#> 7              C1       C1         N/A
#> 3          C2-CGP       C2         CGP
#> 15          C2-CP       C2          CP
#> 5  C2-CP:BIOCARTA       C2 CP:BIOCARTA
#> 13     C2-CP:KEGG       C2     CP:KEGG
#> 16      C2-CP:PID       C2      CP:PID
#> 17 C2-CP:REACTOME       C2 CP:REACTOME
#> 1          C3-MIR       C3         MIR
#> 2          C3-TFT       C3         TFT
#> 6          C4-CGN       C4         CGN
#> 14          C4-CM       C4          CM
#> 9           C5-BP       C5          BP
#> 10          C5-CC       C5          CC
#> 8           C5-MF       C5          MF
#> 4              C6       C6         N/A
#> 11             C7       C7         N/A
#> 12              H        H         N/A
#>                                                                                                                                                                                                                                                                                                                  Category_Description
#> 7                                                                                                                                                                                                                                                          Gene sets corresponding to each human chromosome and each cytogenetic band
#> 3                                                                                                                                       Gene sets in this collection are curated from various sources, including online pathway databases and the biomedical literature. Many sets are also contributed by individual domain experts.
#> 15                                                                                                                                      Gene sets in this collection are curated from various sources, including online pathway databases and the biomedical literature. Many sets are also contributed by individual domain experts.
#> 5                                                                                                                                       Gene sets in this collection are curated from various sources, including online pathway databases and the biomedical literature. Many sets are also contributed by individual domain experts.
#> 13                                                                                                                                      Gene sets in this collection are curated from various sources, including online pathway databases and the biomedical literature. Many sets are also contributed by individual domain experts.
#> 16                                                                                                                                      Gene sets in this collection are curated from various sources, including online pathway databases and the biomedical literature. Many sets are also contributed by individual domain experts.
#> 17                                                                                                                                      Gene sets in this collection are curated from various sources, including online pathway databases and the biomedical literature. Many sets are also contributed by individual domain experts.
#> 1                                                Gene sets representing potential targets of regulation by transcription factors or microRNAs. The sets consist of genes grouped by elements they share in their non-protein coding regions. The elements represent known or likely cis-regulatory elements in promoters and 3'-UTRs.
#> 2                                                Gene sets representing potential targets of regulation by transcription factors or microRNAs. The sets consist of genes grouped by elements they share in their non-protein coding regions. The elements represent known or likely cis-regulatory elements in promoters and 3'-UTRs.
#> 6                                                                                                                                                                                                                                     Computational gene sets defined by mining large collections of cancer-oriented microarray data.
#> 14                                                                                                                                                                                                                                    Computational gene sets defined by mining large collections of cancer-oriented microarray data.
#> 9                                                                                                                                                                                                                                                                         Gene sets that contain genes annotated by the same GO term.
#> 10                                                                                                                                                                                                                                                                        Gene sets that contain genes annotated by the same GO term.
#> 8                                                                                                                                                                                                                                                                         Gene sets that contain genes annotated by the same GO term.
#> 4                                            Gene sets that represent signatures of cellular pathways which are often dis-regulated in cancer. The majority of signatures were generated directly from microarray data from NCBI GEO or from internal unpublished profiling experiments involving perturbation of known cancer genes.
#> 11                                                                                                                                              Gene sets that represent cell states and perturbations within the immune system. The signatures were generated by manual curation of published studies in human and mouse immunology.
#> 12 Hallmark gene sets summarize and represent specific well-defined biological states or processes and display coherent expression. These gene sets were generated by a computational methodology based on identifying overlaps between gene sets in other MSigDB collections and retaining genes that display coordinate expression.
#>                                                                                                                                                                                                                                                                                               Subcategory_Description
#> 7                                                                                                                                                                                                                                                                                           No subcategory available.
#> 3                                                                                                                                                                                                                                                                                  Chemical and genetic perturbations
#> 15                                                                                                                                                                                                                                                                                       Additional currated pathways
#> 5                                                                                                                                                                                                                                            Canonical Pathways gene sets derived from the BioCarta pathway database.
#> 13                                                                                                                                                                                                                                               Canonical Pathways gene sets derived from the KEGG pathway database.
#> 16                                                                                                                                                                                                                 Canonical Pathways gene sets derived from the Pathway Interaction Database (PID) pathway database.
#> 17                                                                                                                                                                                                                                           Canonical Pathways gene sets derived from the Reactome pathway database.
#> 1                                                                                                                                                                                                          All miRNA target prediction gene sets. Combined superset of both miRDB prediction methods and legacy sets.
#> 2                                                                                                                                                                                            All transcription factor target prediction gene sets. Combined superset of both GTRD prediction methods and legacy sets.
#> 6                                                                                                                                                              Gene sets defined by expression neighborhoods centered on 380 cancer-associated genes. This collection is described in Subramanian, Tamayo et al. 2005
#> 14 Gene sets defined by Segal et al. 2004. Briefly, the authors compiled gene sets ('modules') from a variety of resources such as KEGG, GO, and others. By mining a large compendium of cancer-related microarray data, they identified 456 such modules as significantly changed in a variety of cancer conditions.
#> 9                                                                                                                                                                                                                                                          Gene sets derived from the GO Biological Process Ontology.
#> 10                                                                                                                                                                                                                                                         Gene sets derived from the GO Cellular Component Ontology.
#> 8                                                                                                                                                                                                                                                          Gene sets derived from the GO Molecular Function Ontology.
#> 4                                                                                                                                                                                                                                                                                           No subcategory available.
#> 11                                                                                                                                                                                                                                                                                          No subcategory available.
#> 12                                                                                                                                                                                                                                                                                          No subcategory available.