Retrieve example datasets

Retrieves published example datasets stored in SingleCellExperiment using the scRNAseq and TENxPBMCData packages. See 'Details' for a list of available datasets.

importExampleData(
  dataset,
  class = c("Matrix", "matrix"),
  delayedArray = FALSE,
  rowNamesDedup = TRUE
)

Arguments

dataset

Character. Name of the dataset to retrieve.

class

Character. The class of the expression matrix stored in the SCE object. Can be one of "Matrix" or "matrix". "Matrix" will store the data as a sparse matrix from package Matrix while "matrix" will store the data in a standard matrix. Default "Matrix".

delayedArray

Boolean. Whether to read the expression matrix as DelayedArray object or not. Default FALSE.

rowNamesDedup

Boolean. Whether to deduplicate rownames. Default TRUE.

Value

The specified SingleCellExperiment object.

Details

See the list below for the available datasets and their descriptions.

"fluidigm_pollen"

Retrieved with ReprocessedFluidigmData. Returns a dataset of 65 human neural cells from Pollen et al. (2014), each sequenced at high and low coverage (SRA accession SRP041736).

"allen_tasic"

Retrieved with ReprocessedAllenData. Returns a dataset of 379 mouse brain cells from Tasic et al. (2016).

"pbmc3k"

Retrieved with TENxPBMCData. 2,700 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc4k"

Retrieved with TENxPBMCData. 4,340 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc6k"

Retrieved with TENxPBMCData. 5,419 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc8k"

Retrieved with TENxPBMCData. 8,381 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc33k"

Retrieved with TENxPBMCData. 33,148 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

"pbmc68k"

Retrieved with TENxPBMCData. 68,579 peripheral blood mononuclear cells (PBMCs) from 10X Genomics.

Author

Joshua D. Campbell, David Jenkins

Examples

sce <- importExampleData("pbmc3k")
#> snapshotDate(): 2021-10-19
#> see ?TENxPBMCData and browseVignettes('TENxPBMCData') for documentation
#> Warning: Corrupt Cache: resource path
#>   See AnnotationHub's TroubleshootingTheCache vignette section on corrupt cache
#>   cache: C:\Users\Yichen\AppData\Local/R/cache/R/ExperimentHub
#>   potential duplicate files: 
#>     66c41363c8_1605
#>     b1581a4f3a4c_1605
#>     66cb6b3970_1606
#>     b15821a933a6_1606
#>     66c4fe3642_1607
#>     b158265122b6_1607
#>     28582f9d3cbf_3508
#>     5e4c7beb372e_3508
#>     285859427542_3509
#>     5e4c531d2b58_3509
#> Continuing with first found cached file
#> downloading 1 resources
#> retrieving 1 resource
#> loading from cache
#> Warning: Corrupt Cache: resource path
#>   See AnnotationHub's TroubleshootingTheCache vignette section on corrupt cache
#>   cache: C:\Users\Yichen\AppData\Local/R/cache/R/ExperimentHub
#>   potential duplicate files: 
#>     66c41363c8_1605
#>     b1581a4f3a4c_1605
#>     66cb6b3970_1606
#>     b15821a933a6_1606
#>     66c4fe3642_1607
#>     b158265122b6_1607
#>     28582f9d3cbf_3508
#>     5e4c7beb372e_3508
#>     285859427542_3509
#>     5e4c531d2b58_3509
#> Continuing with first found cached file
#> Warning: Corrupt Cache: resource path
#>   See AnnotationHub's TroubleshootingTheCache vignette section on corrupt cache
#>   cache: C:\Users\Yichen\AppData\Local/R/cache/R/ExperimentHub
#>   potential duplicate files: 
#>     66c41363c8_1605
#>     b1581a4f3a4c_1605
#>     66cb6b3970_1606
#>     b15821a933a6_1606
#>     66c4fe3642_1607
#>     b158265122b6_1607
#>     28582f9d3cbf_3508
#>     5e4c7beb372e_3508
#>     285859427542_3509
#>     5e4c531d2b58_3509
#> Continuing with first found cached file
#> Warning: Corrupt Cache: resource path
#>   See AnnotationHub's TroubleshootingTheCache vignette section on corrupt cache
#>   cache: C:\Users\Yichen\AppData\Local/R/cache/R/ExperimentHub
#>   potential duplicate files: 
#>     66c41363c8_1605
#>     b1581a4f3a4c_1605
#>     66cb6b3970_1606
#>     b15821a933a6_1606
#>     66c4fe3642_1607
#>     b158265122b6_1607
#>     28582f9d3cbf_3508
#>     5e4c7beb372e_3508
#>     285859427542_3509
#>     5e4c531d2b58_3509
#> Continuing with first found cached file
#> Warning: Corrupt Cache: resource path
#>   See AnnotationHub's TroubleshootingTheCache vignette section on corrupt cache
#>   cache: C:\Users\Yichen\AppData\Local/R/cache/R/ExperimentHub
#>   potential duplicate files: 
#>     66c41363c8_1605
#>     b1581a4f3a4c_1605
#>     66cb6b3970_1606
#>     b15821a933a6_1606
#>     66c4fe3642_1607
#>     b158265122b6_1607
#>     28582f9d3cbf_3508
#>     5e4c7beb372e_3508
#>     285859427542_3509
#>     5e4c531d2b58_3509
#> Continuing with first found cached file
#> Warning: Corrupt Cache: resource path
#>   See AnnotationHub's TroubleshootingTheCache vignette section on corrupt cache
#>   cache: C:\Users\Yichen\AppData\Local/R/cache/R/ExperimentHub
#>   potential duplicate files: 
#>     66c41363c8_1605
#>     b1581a4f3a4c_1605
#>     66cb6b3970_1606
#>     b15821a933a6_1606
#>     66c4fe3642_1607
#>     b158265122b6_1607
#>     28582f9d3cbf_3508
#>     5e4c7beb372e_3508
#>     285859427542_3509
#>     5e4c531d2b58_3509
#> Continuing with first found cached file