Plot a heatmap to visualize the result of findMarkerDiffExp

Plot a heatmap to visualize the result of findMarkerDiffExp

plotMarkerDiffExp(
  inSCE,
  useAssay = "logcounts",
  orderBy = "size",
  log2fcThreshold = 1,
  fdrThreshold = 0.05,
  decreasing = TRUE,
  rowDataName = NULL,
  colDataName = NULL,
  featureAnnotations = NULL,
  cellAnnotations = NULL,
  featureAnnotationColor = NULL,
  cellAnnotationColor = NULL,
  rowSplitBy = ifelse(is.null(orderBy), NULL, colnames(inSCE@metadata$findMarker)[5]),
  colSplitBy = rowSplitBy,
  ...
)

Arguments

inSCE
SingleCellExperiment inherited object.

useAssay
character. A string specifying which assay to use for the expression values. Default "logcounts".

orderBy
The ordering method of the clusters on the splitted heatmap. Can be chosen from "size" or "name", specified with vector of ordered unique cluster labels, or set as NULL for unsplitted heatmap. Default "size".

log2fcThreshold
Only use DEGs with the absolute values of log2FC larger than this value. Default 1

fdrThreshold
Only use DEGs with FDR value smaller than this value. Default 0.05

decreasing
Order the cluster decreasingly. Default TRUE.

rowDataName
character. The column name(s) in rowData that need to be added to the annotation. Default NULL.

colDataName
character. The column name(s) in colData that need to be added to the annotation. Default NULL.

featureAnnotations
data.frame, with rownames containing all the features going to be plotted. Character columns should be factors. Default NULL.

cellAnnotations
data.frame, with rownames containing all the cells going to be plotted. Character columns should be factors. Default NULL.

featureAnnotationColor
A named list. Customized color settings for feature labeling. Should match the entries in the featureAnnotations or rowDataName. For each entry, there should be a list/vector of colors named with categories. Default NULL.

cellAnnotationColor
A named list. Customized color settings for cell labeling. Should match the entries in the cellAnnotations or colDataName. For each entry, there should be a list/vector of colors named with categories. Default NULL.

rowSplitBy
character. Do semi-heatmap based on the grouping of this(these) annotation(s). Should exist in either rowDataName or names(featureAnnotations). Default is the value of cluster in findMarkerDiffExp when orderBy is not NULL, or NULL otherwise.

colSplitBy
character. Do semi-heatmap based on the grouping of this(these) annotation(s). Should exist in either colDataName or names(cellAnnotations). Default is the same as rowSplitBy.

...
Other arguments passed to plotSCEHeatmap.

Value

A Heatmap object