Perform comprehensive single cell QC

A wrapper function to run several QC algorithms on a SingleCellExperiment object containing cells after empty droplets have been removed.

runCellQC(
  inSCE,
  algorithms = c("QCMetrics", "doubletCells", "cxds", "bcds", "cxds_bcds_hybrid",
    "scrublet", "doubletFinder", "decontX"),
  sample = NULL,
  collectionName = NULL,
  geneSetList = NULL,
  geneSetListLocation = "rownames",
  geneSetCollection = NULL,
  useAssay = "counts",
  seed = 12345,
  paramsList = NULL
)

Arguments

inSCE
A SingleCellExperiment object.

algorithms
Character vector. Specify which QC algorithms to run. Available options are "QCMetrics", "scrublet", "doubletCells", "cxds", "bcds", "cxds_bcds_hybrid", and "decontX".

sample
Character vector. Indicates which sample each cell belongs to. Algorithms will be run on cells from each sample separately.

collectionName
Character. Name of a GeneSetCollection obtained by using one of the importGeneSet* functions. Default NULL.

geneSetList
See runPerCellQC. Default NULL.

geneSetListLocation
See runPerCellQC. Default NULL.

geneSetCollection
See runPerCellQC. Default NULL.

useAssay
A string specifying which assay contains the count matrix for cells.

seed
Seed for the random number generator. Default 12345.

paramsList
A list containing parameters for QC functions. Default NULL.

Value

SingleCellExperiment object containing the outputs of the specified algorithms in the colData of inSCE.

Examples

# \donttest{
data(scExample, package = "singleCellTK")
sce <- sce[, colData(sce)$type != 'EmptyDroplet']
sce <- runCellQC(sce)

#> Wed Oct 14 12:01:27 2020 ... Running 'perCellQCMetrics'

#> Wed Oct 14 12:01:27 2020 ... Running 'scrublet'

#> Wed Oct 14 12:01:29 2020 ... Running 'doubletCells'

#> Wed Oct 14 12:01:30 2020 ... Running 'doubletFinder'

#> Wed Oct 14 12:01:41 2020 ... Running 'cxds'

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Wed Oct 14 12:01:41 2020 ... Running 'bcds'

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Wed Oct 14 12:01:42 2020 ... Running 'cxds_bcds_hybrid'

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Wed Oct 14 12:01:43 2020 ... Running 'DecontX'

#> --------------------------------------------------

#> Starting DecontX

#> --------------------------------------------------

#> Wed Oct 14 12:01:43 2020 .. Analyzing all cells

#> Wed Oct 14 12:01:43 2020 .... Generating UMAP and estimating cell types

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Wed Oct 14 12:01:45 2020 .... Estimating contamination

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 10 | converge: 0.0278

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 20 | converge: 0.01058

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 30 | converge: 0.00584

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 40 | converge: 0.004697

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 50 | converge: 0.003356

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 60 | converge: 0.002634

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 70 | converge: 0.001941

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 80 | converge: 0.001413

#> Wed Oct 14 12:01:45 2020 ...... Completed iteration: 90 | converge: 0.001284

#> Wed Oct 14 12:01:46 2020 ...... Completed iteration: 100 | converge: 0.00111

#> Wed Oct 14 12:01:46 2020 ...... Completed iteration: 104 | converge: 0.0009827

#> Wed Oct 14 12:01:46 2020 .. Finalizing decontaminated matrix

#> --------------------------------------------------

#> Completed DecontX. Total time: 2.605954 secs

#> --------------------------------------------------

# }