Perform comprehensive single cell QC

A wrapper function to run several QC algorithms on a SingleCellExperiment object containing cells after empty droplets have been removed.

runCellQC(
  inSCE,
  algorithms = c("QCMetrics", "scDblFinder", "cxds", "bcds", "cxds_bcds_hybrid",
    "decontX", "decontX_bg", "soupX", "soupX_bg"),
  sample = NULL,
  collectionName = NULL,
  geneSetList = NULL,
  geneSetListLocation = "rownames",
  geneSetCollection = NULL,
  mitoRef = "human",
  mitoIDType = "ensembl",
  mitoPrefix = "MT-",
  mitoID = NULL,
  mitoGeneLocation = "rownames",
  useAssay = "counts",
  background = NULL,
  bgAssayName = NULL,
  bgBatch = NULL,
  seed = 12345,
  paramsList = NULL
)

Arguments

inSCE

A SingleCellExperiment object.

algorithms

Character vector. Specify which QC algorithms to run. Available options are "QCMetrics", "scrublet", "doubletFinder", "scDblFinder", "cxds", "bcds", "cxds_bcds_hybrid", "decontX" and "soupX".

sample

Character vector. Indicates which sample each cell belongs to. Algorithms will be run on cells from each sample separately.

collectionName

Character. Name of a GeneSetCollection obtained by using one of the importGeneSet* functions. Default NULL.

geneSetList

See runPerCellQC. Default NULL.

geneSetListLocation

See runPerCellQC. Default NULL.

geneSetCollection

See runPerCellQC. Default NULL.

mitoRef, mitoIDType, mitoPrefix, mitoID, mitoGeneLocation

Arguments used to import mitochondrial genes and quantify their expression. Please see runPerCellQC for detailed information.

useAssay

A string specifying which assay contains the count matrix for cells.

background

A SingleCellExperiment with the matrix located in the assay slot under bgAssayName. It should have the same structure as inSCE except it contains the matrix of empty droplets instead of cells. When supplied, empirical distribution of transcripts from these empty droplets will be used as the contamination distribution. It is only used in algorithms "decontX" and "soupX". Default NULL.

bgAssayName

Character. Name of the assay to use if background is a SingleCellExperiment. If NULL, the function will use the same value as useAssay. It is only used in algorithms "decontX" and "soupX". Default is NULL.

bgBatch

Batch labels for background. If background is a SingleCellExperiment object, this can be a single character specifying a name that can be found in colData(background) to directly use the barcode annotation Its unique values should be the same as those in sample, such that each batch of cells have their corresponding batch of empty droplets as background, pointed by this parameter. It is only used in algorithms "decontX" and "soupX". Default to NULL.

seed

Seed for the random number generator. Default 12345.

paramsList

A list containing parameters for QC functions. Default NULL.

Value

SingleCellExperiment object containing the outputs of the specified algorithms in the colData of inSCE.

Examples

data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
if (FALSE) { # \dontrun{
sce <- runCellQC(sce)
} # }