Perform comprehensive single cell QC

A wrapper function to run several QC algorithms on a SingleCellExperiment object containing cells after empty droplets have been removed.

runCellQC(
  inSCE,
  algorithms = c("QCMetrics", "doubletCells", "cxds", "bcds", "cxds_bcds_hybrid",
    "scrublet", "doubletFinder", "decontX"),
  sample = NULL,
  geneSetList = NULL,
  geneSetListLocation = "rownames",
  geneSetCollection = NULL,
  useAssay = "counts",
  seed = 12345,
  paramsList = NULL
)

Arguments

inSCE
A SingleCellExperiment object.

algorithms
Character vector. Specify which QC algorithms to run. Available options are "QCMetrics", "scrublet", "doubletCells", "cxds", "bcds", "cxds_bcds_hybrid", and "decontX".

sample
Character vector. Indicates which sample each cell belongs to. Algorithms will be run on cells from each sample separately.

geneSetList
See runPerCellQC. Default NULL.

geneSetListLocation
See runPerCellQC. Default NULL.

geneSetCollection
See runPerCellQC. Default NULL.

useAssay
A string specifying which assay contains the count matrix for cells.

seed
Seed for the random number generator. Default 12345.

paramsList
A list containing parameters for QC functions. Default NULL.

Value

SingleCellExperiment object containing the outputs of the specified algorithms in the colData of inSCE.

Examples

# \donttest{
data(scExample, package = "singleCellTK")
sce <- sce[, colData(sce)$type != 'EmptyDroplet']
sce <- runCellQC(sce)

#> Tue Jun 30 16:57:36 2020 ... Running 'perCellQCMetrics'

#> Tue Jun 30 16:57:36 2020 ... Running 'scrublet'

#> Tue Jun 30 16:57:39 2020 ... Running 'doubletCells'

#> Tue Jun 30 16:57:41 2020 ... Running 'doubletFinder'

#> Tue Jun 30 16:57:55 2020 ... Running 'cxds'

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Tue Jun 30 16:57:56 2020 ... Running 'bcds'

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Tue Jun 30 16:57:58 2020 ... Running 'cxds_bcds_hybrid'

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Setting levels: control = 0, case = 1

#> Setting direction: controls < cases

#> Tue Jun 30 16:57:59 2020 ... Running 'DecontX'

#> --------------------------------------------------

#> Starting DecontX

#> --------------------------------------------------

#> Tue Jun 30 16:57:59 2020 .. Analyzing all cells

#> Tue Jun 30 16:57:59 2020 .... Generating UMAP and estimating cell types

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Tue Jun 30 16:58:03 2020 .... Estimating contamination

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 10 | converge: 0.0278

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 20 | converge: 0.01058

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 30 | converge: 0.00584

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 40 | converge: 0.004697

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 50 | converge: 0.003356

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 60 | converge: 0.002634

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 70 | converge: 0.001941

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 80 | converge: 0.001413

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 90 | converge: 0.001284

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 100 | converge: 0.00111

#> Tue Jun 30 16:58:03 2020 ...... Completed iteration: 104 | converge: 0.0009827

#> Tue Jun 30 16:58:03 2020 .. Finalizing decontaminated matrix

#> --------------------------------------------------

#> Completed DecontX. Total time: 4.437452 secs

#> --------------------------------------------------

# }