Find doublets/multiplets using cxds.

A wrapper function for cxds. Annotate doublets/multiplets using co-expression based approach. Generate a doublet score for each cell. Infer doublets if estNdbl is TRUE.

runCxds(
  inSCE,
  sample = NULL,
  seed = 12345,
  ntop = 500,
  binThresh = 0,
  verb = FALSE,
  retRes = FALSE,
  estNdbl = FALSE
)

Arguments

inSCE
A SingleCellExperiment object. Needs counts in assays slot.

sample
Character vector. Indicates which sample each cell belongs to. cxds will be run on cells from each sample separately. If NULL, then all cells will be processed together. Default NULL.

seed
Seed for the random number generator. Default 12345.

ntop
See cxds for more information. Default 500.

binThresh
See cxds for more information. Default 0.

verb
See cxds for more information. Default FALSE.

retRes
See cxds for more information. Default FALSE.

estNdbl
See cxds for more information. Default FALSE.

Value

A SingleCellExperiment object with cxds output appended to the colData slot. The columns include cxds_score and optionally cxds_call. Please refer to the documentation of cxds for details.

Examples

data(scExample, package = "singleCellTK")
sce <- sce[, colData(sce)$type != 'EmptyDroplet']
sce <- runCxds(sce)

#> Tue Jun 30 16:58:04 2020 ... Running 'cxds'