Detecting contamination with DecontX.

A wrapper function for decontX. Identify potential contamination from experimental factors such as ambient RNA.

runDecontX(
  inSCE,
  sample = NULL,
  useAssay = "counts",
  background = NULL,
  bgAssayName = NULL,
  bgBatch = NULL,
  z = NULL,
  maxIter = 500,
  delta = c(10, 10),
  estimateDelta = TRUE,
  convergence = 0.001,
  iterLogLik = 10,
  varGenes = 5000,
  dbscanEps = 1,
  seed = 12345,
  logfile = NULL,
  verbose = TRUE
)

Arguments

inSCE

A SingleCellExperiment object.

sample

A single character specifying a name that can be found in colData(inSCE) to directly use the cell annotation; or a character vector with as many elements as cells to indicates which sample each cell belongs to. Default NULL. decontX will be run on cells from each sample separately.

useAssay

A string specifying which assay in the SCE to use. Default 'counts'.

background

A SingleCellExperiment with the matrix located in the assay slot under bgAssayName. It should have the same structure as inSCE except it contains the matrix of empty droplets instead of cells. When supplied, empirical distribution of transcripts from these empty droplets will be used as the contamination distribution. Default NULL.

bgAssayName

Character. Name of the assay to use if background is a SingleCellExperiment. If NULL, the function will use the same value as useAssay. Default is NULL.

bgBatch

Batch labels for background. If background is a SingleCellExperiment object, this can be a single character specifying a name that can be found in colData(background) to directly use the barcode annotation; or a numeric / character vector that has as many elements as barcodes to indicate which sample each barcode belongs to. Its unique values should be the same as those in sample, such that each batch of cells have their corresponding batch of empty droplets as background, pointed by this parameter. Default to NULL.

z

Numeric or character vector. Cell cluster labels. If NULL, PCA will be used to reduce the dimensionality of the dataset initially, 'umap' from the 'uwot' package will be used to further reduce the dataset to 2 dimenions and the 'dbscan' function from the 'dbscan' package will be used to identify clusters of broad cell types. Default NULL.

maxIter

Integer. Maximum iterations of the EM algorithm. Default 500.

delta

Numeric Vector of length 2. Concentration parameters for the Dirichlet prior for the contamination in each cell. The first element is the prior for the native counts while the second element is the prior for the contamination counts. These essentially act as pseudocounts for the native and contamination in each cell. If estimateDelta = TRUE, this is only used to produce a random sample of proportions for an initial value of contamination in each cell. Then fit_dirichlet is used to update delta in each iteration. If estimateDelta = FALSE, then delta is fixed with these values for the entire inference procedure. Fixing delta and setting a high number in the second element will force decontX to be more aggressive and estimate higher levels of contamination at the expense of potentially removing native expression. Default c(10, 10).

estimateDelta

Boolean. Whether to update delta at each iteration.

convergence

Numeric. The EM algorithm will be stopped if the maximum difference in the contamination estimates between the previous and current iterations is less than this. Default 0.001.

iterLogLik

Integer. Calculate log likelihood every iterLogLik iteration. Default 10.

varGenes

Integer. The number of variable genes to use in dimensionality reduction before clustering. Variability is calcualted using modelGeneVar function from the 'scran' package. Used only when z is not provided. Default 5000.

dbscanEps

Numeric. The clustering resolution parameter used in 'dbscan' to estimate broad cell clusters. Used only when z is not provided. Default 1.

seed

Integer. Passed to with_seed. For reproducibility, a default value of 12345 is used. If NULL, no calls to with_seed are made.

logfile

Character. Messages will be redirected to a file named `logfile`. If NULL, messages will be printed to stdout. Default NULL.

verbose

Logical. Whether to print log messages. Default TRUE.

Value

A SingleCellExperiment object with 'decontX_Contamination' and 'decontX_Clusters' added to the colData slot. Additionally, the decontaminated counts will be added as an assay called 'decontXCounts'.

Examples

data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runDecontX(sce[,sample(ncol(sce),20)])
#> Wed Apr 23 11:40:38 2025 ... Running 'DecontX'
#> --------------------------------------------------
#> Starting DecontX
#> --------------------------------------------------
#> Wed Apr 23 11:40:38 2025 .. Analyzing all cells
#> Wed Apr 23 11:40:39 2025 .... Generating UMAP and estimating cell types
#> Wed Apr 23 11:40:42 2025 .... Estimating contamination
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 10 | converge: 0.01985
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 20 | converge: 0.01116
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 30 | converge: 0.006593
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 40 | converge: 0.003919
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 50 | converge: 0.002691
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 60 | converge: 0.002093
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 70 | converge: 0.001932
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 80 | converge: 0.001769
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 90 | converge: 0.001605
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 100 | converge: 0.00145
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 110 | converge: 0.001304
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 120 | converge: 0.001168
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 130 | converge: 0.001041
#> Wed Apr 23 11:40:42 2025 ...... Completed iteration: 134 | converge: 0.0009927
#> Wed Apr 23 11:40:42 2025 .. Calculating final decontaminated matrix
#> --------------------------------------------------
#> Completed DecontX. Total time: 3.68077 secs
#> --------------------------------------------------