Detecting contamination with DecontX.

A wrapper function for decontX. Identify potential contamination from experimental factors such as ambient RNA.

runDecontX(inSCE, sample = NULL, useAssay = "counts", ...)

Arguments

inSCE
A SingleCellExperiment object.

sample
Character vector. Indicates which sample each cell belongs to. Default NULL. decontX will be run on cells from each sample separately.

useAssay
A string specifying which assay in the SCE to use. Default 'counts'.

...
Additional arguments to pass to decontX.

Value

A SingleCellExperiment object with 'decontX_Contamination' and 'decontX_Clusters' added to the colData slot. Additionally, the decontaminated counts will be added as an assay called 'decontXCounts'.

Examples

data(scExample, package = "singleCellTK")
sce <- sce[, colData(sce)$type != 'EmptyDroplet']
sce <- runDecontX(sce)

#> Tue Jun 30 16:58:08 2020 ... Running 'DecontX'

#> --------------------------------------------------

#> Starting DecontX

#> --------------------------------------------------

#> Tue Jun 30 16:58:08 2020 .. Analyzing all cells

#> Tue Jun 30 16:58:08 2020 .... Generating UMAP and estimating cell types

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Found more than one class "dist" in cache; using the first, from namespace 'BiocGenerics'

#> Also defined by 'spam'

#> Tue Jun 30 16:58:12 2020 .... Estimating contamination

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 10 | converge: 0.0278

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 20 | converge: 0.01058

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 30 | converge: 0.00584

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 40 | converge: 0.004697

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 50 | converge: 0.003356

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 60 | converge: 0.002634

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 70 | converge: 0.001941

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 80 | converge: 0.001413

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 90 | converge: 0.001284

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 100 | converge: 0.00111

#> Tue Jun 30 16:58:12 2020 ...... Completed iteration: 104 | converge: 0.0009827

#> Tue Jun 30 16:58:12 2020 .. Finalizing decontaminated matrix

#> --------------------------------------------------

#> Completed DecontX. Total time: 4.254559 secs

#> --------------------------------------------------