Detect doublet cells using doubletCells.

A wrapper function for doubletCells. Identify potential doublet cells based on simulations of putative doublet expression profiles. Generate a doublet score for each cell.

runDoubletCells(
  inSCE,
  sample = NULL,
  useAssay = "counts",
  nNeighbors = 50,
  simDoublets = 10000,
  seed = 12345,
  ...
)

Arguments

inSCE
A SingleCellExperiment object.

sample
Character vector. Indicates which sample each cell belongs to. doubletCells will be run on cells from each sample separately.

useAssay
A string specifying which assay in the SCE to use.

nNeighbors
Number of nearest neighbors used to calculate density for doublet detection. Default 50.

simDoublets
Number of simulated doublets created for doublet detection. Default 10000.

seed
Seed for the random number generator. Default 12345.

...
Additional arguments to pass to doubletCells.

Value

A SingleCellExperiment object with the 'scran_doubletCell_Score' column added to the colData slot.

Details

This function is a wrapper function for doubletCells. runDoubletCells runs doubletCells for each sample within inSCE iteratively. The resulting doublet scores for all cells will be appended to the colData of inSCE.

References

Lun ATL (2018). Detecting doublet cells with scran. https://ltla.github.io/SingleCellThoughts/software/doublet_detectio/bycell.html

See also

Examples

data(scExample, package = "singleCellTK")
sce <- sce[, colData(sce)$type != 'EmptyDroplet']
sce <- runDoubletCells(sce)

#> Tue Jun 30 16:58:13 2020 ... Running 'doubletCells'