Generates a doublet score for each cell via doubletFinder

Uses doubletFinder to determine cells within the dataset suspected to be doublets.

runDoubletFinder(
  inSCE,
  useAssay = "counts",
  sample = NULL,
  seed = 12345,
  seuratNfeatures = 2000,
  seuratPcs = seq(15),
  seuratRes = 1.5,
  formationRate = 0.075,
  nCores = NULL,
  verbose = FALSE
)

Arguments

inSCE

Input SingleCellExperiment object. Must contain a counts matrix

useAssay

Indicate which assay to use. Default "counts".

sample

Numeric vector. Each cell will be assigned a sample number.

seed

Seed for the random number generator. Default 12345.

seuratNfeatures

Integer. Number of highly variable genes to use. Default 2000.

seuratPcs

Numeric vector. The PCs used in seurat function to determine number of clusters. Default 1:15.

seuratRes

Numeric vector. The resolution parameter used in seurat, which adjusts the number of clusters determined via the algorithm. Default 1.5.

formationRate

Doublet formation rate used within algorithm. Default 0.075.

nCores

Number of cores used for running the function.

verbose

Boolean. Wheter to print messages from Seurat and DoubletFinder. Default FALSE.

Value

SingleCellExperiment object containing the 'doublet_finder_doublet_score'.

Examples

data(scExample, package = "singleCellTK")
sce <- subsetSCECols(sce, colData = "type != 'EmptyDroplet'")
sce <- runDoubletFinder(sce)
#> Thu Apr 28 11:27:56 2022 ... Running 'doubletFinder'